### 2024-01-29 #6300 @editorialbot
[PRE REVIEW]: rfsed: Receiver function analysis and dealing with sediment effects
**Submitting author:** @akinremisa (<a href="http://orcid.org/0000-0001-9162-7156">Stephen Akinremi</a>)
**Repository:** https://github.com/akinremisa/rfsed
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.7
