### 2024-01-29 #6300 @kthyng
Hi @akinremisa and thanks for your submission! We have a backlog of submissions so I will add this to our waitlist. In the meantime, please suggest 5 reviewers from the database listed above or your own (non-conflicted) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them. Also please check out the other notifications from the editorial bot above. I see that you have very light docs that reference examples but the examples are all scripts — typically there would be notebooks or a hosted readthedocs version of examples/tutorials. Have you considered that sort of thing?

Thanks for your patience.
