### 2024-01-29 #6300 @akinremisa
Hi @kthyng
Thank you for initiating the review of the manuscript and software I submitted.

Is it compulsory to add a section on "Statement of need" in the paper?
