### 2024-01-29 #6300 @kthyng
> Is it compulsory to add a section on "Statement of need" in the paper?

Yes: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain
