### 2024-01-29 518b83d
Changed example files from script to notebooks
