+++ joss/paper.md
@@ -30,8 +30,7 @@ bibliography: paper.bib
 The receiver function technique is a well-established technique to image velocity contrasts in the subsurface (such as the crust-mantle boundary - Moho) by using isolated P–to-S wave conversions (P-receiver functions) and the reverberations generated at such discontinuities. Several methods such as H-k stacking [@Zhu2000], have been developed to investigate the average crustal thickness and Vp/Vs ratio using receiver functions. However, the presence of a near-surface low-velocity sedimentary layer can obscure Moho phases due to the additional P-to-S wave conversions and associated reverberated phases created at the sediment-basement discontinuity. These additional intra-crustal phases can have large amplitudes and similar arrival times as the Moho phases, which makes it difficult to retrieve Moho information using standard RF techniques. 
 
 # Statement of need
-``rfsed`` is a Python package for receiver function analysis that includes methods for dealing with sediment effects. ``rfsed`` presents a new approach to retrieving the sediment thickness properties using H-K stacking of the high-frequency receiver function, followed by a waveform fitting approach to retrieve the average crustal thickness and Vp/Vs [@Akinremi2024]. Moreover, ``rfsed`` contains implementations of the most common receiver function approaches for dealing with the sediment effect with possible synthetic testing capabilities. ``rfsed`` has already been used in [@Akinremi2024] for multimethod imaging of the crustal structure using receiver functions obtained from stations overlying seimentary layers imaging the crustal structure 
-
+``rfsed`` is a Python package for receiver function analysis that includes methods for dealing with sediment effects. ``rfsed`` presents a new approach to retrieving the sediment thickness properties using H-K stacking of the high-frequency receiver function, followed by a waveform fitting approach to retrieve the average crustal thickness and Vp/Vs [@Akinremi2024]. Moreover, ``rfsed`` contains implementations of the most common receiver function approaches for dealing with the sediment effect with possible synthetic testing capabilities. ``rfsed`` has already been used in [@Akinremi2024] for multimethod imaging of the crustal structure using receiver functions obtained from stations overlying seimentary layer.
 # Key Functionality
 ``rfsed`` contains modules to carry out H-k stacking [@Zhu2000], sequential H-k stacking [@Yeck2013], resonance filtering and modified H-k stacking [@Yu2015], and waveform Fitting [@Akinremi2024] with possible synthetic waveform generation for 1D earth models based on [@Xu2022] to test of the previously mentioned methods. It comes with tools to create high-quality figures, which include result plots for H-k stacking, sequential H-k stacking (e.g., Figure 1), resonance filtering, and waveform-fitting methods (e.g., Figure 2). Besides these methods, ``rfsed`` has modules for extracting earthquake waveforms from local seismic record files. There are multiprocessing options for waveform fitting and extracting earthquake data from local seismic record files for higher efficiency.
 
