+++ joss/paper.md
@@ -31,6 +31,7 @@ The receiver function technique is a well-established technique to image velocit
 
 # Statement of need
 ``rfsed`` is a Python package for receiver function analysis that includes methods for dealing with sediment effects. ``rfsed`` presents a new approach to retrieving the sediment thickness properties using H-K stacking of the high-frequency receiver function, followed by a waveform fitting approach to retrieve the average crustal thickness and Vp/Vs [@Akinremi2024]. Moreover, ``rfsed`` contains implementations of the most common receiver function approaches for dealing with the sediment effect with possible synthetic testing capabilities. ``rfsed`` has already been used in [@Akinremi2024] for multimethod imaging of the crustal structure using receiver functions obtained from stations overlying seimentary layer.
+
 # Key Functionality
 ``rfsed`` contains modules to carry out H-k stacking [@Zhu2000], sequential H-k stacking [@Yeck2013], resonance filtering and modified H-k stacking [@Yu2015], and waveform Fitting [@Akinremi2024] with possible synthetic waveform generation for 1D earth models based on [@Xu2022] to test of the previously mentioned methods. It comes with tools to create high-quality figures, which include result plots for H-k stacking, sequential H-k stacking (e.g., Figure 1), resonance filtering, and waveform-fitting methods (e.g., Figure 2). Besides these methods, ``rfsed`` has modules for extracting earthquake waveforms from local seismic record files. There are multiprocessing options for waveform fitting and extracting earthquake data from local seismic record files for higher efficiency.
 
