### 2024-02-05 #6300 @kthyng
@akinremisa — notebooks are not a requirement for your docs for JOSS but I suggested it since typically with notebooks or markdown in hosted docs, users are better able to understand, set up, and use your software properly. It looks like you dumped the contents of the .py files into .ipynb files but there is no explanation of how to use them or what they do, unless I missed some overview that does explain it all. In any case, reviewers will dig deeper into your repository to see what changes should be made and any changes could be made at that point instead.
