### 2024-02-06 2c9e883
Changed calculation of time to include pre-onset time used in the receiver function
