### 2024-02-08 a4544f3
Added 'onset' as stats attribute to waveforms extraceted from local seismic data record
