### 2024-02-14 86b3688
Updated the implementation of resonance filter
