+++ src/rfsed/ReverbFilter.py
@@ -5,6 +5,9 @@
 """
 
 import numpy as np
+from scipy.signal import detrend
+from scipy.fft import fft, ifft
+from scipy.signal import find_peaks
 from scipy import signal
 from scipy.signal import argrelextrema, argrelmin, argrelmax
 import math
@@ -12,7 +15,6 @@ import cmath
 import matplotlib.pyplot as plt
 import matplotlib.gridspec as gridspec
 import matplotlib.transforms as mtransforms
-from scipy.fftpack import fft, ifft
 from matplotlib.legend_handler import HandlerLine2D as HL
 import seaborn
 seaborn.set_style("darkgrid", {"axes.facecolor": ".85"})
@@ -52,66 +54,46 @@ def Resonance_Filt(rfstream, preonset):
     l = len(rfstream[0].data)
     t = np.arange(0, l)
     t = (delta *  t) - preonset 
-    Dt=rfstream[0].stats.delta
+    time=t
+    dt=rfstream[0].stats.delta
     StackedRF=rfstream.stack()
-    StackedData=StackedRF[0].data
-    tzero=np.where(t==0)[0]
-    tzero=int(tzero)  
-    start=tzero
-    rf=StackedData[start::] 
-    time = t[start::]
-    Dt=delta
-    N=len(rf)
-    fmax=1/(2*Dt)
-    df=fmax/(N/2)
+    StackedData=StackedRF[0].data 
     #---------------------------------------------
     # Build the frequency vector
-    N2=N/2
     #---------------------------------------------
-    column1=np.arange(0,N2+1)    
-    column2=np.arange(-N2+1,-1) 
-    #---------------------------------------------
-    # Join the two vectors
-    full_column = np.concatenate((column1,column2))
-    #---------------------------------------------
-    f=df*full_column
-    Nf = N/2+1
-    dw=2*np.pi*df
-    w=dw*full_column
-    #----------------------------------------------------------------------------------------------------------------------
-    # Do Auto-correlation
-    #----------------------------------------------------------------------------------------------------------------------
-    D=rf
-    D=D-np.mean(D)
-    D=signal.detrend(D, axis=-1, type='linear')
-    #------------------------
-    autoc=signal.correlate(D,D,mode='full')#,method='auto')
-    autoc=autoc/max(autoc)
-    autoc=autoc[N-1:2*N-1]
-    #----------------------------------------------------------------------------------------------------------------------
-    # Find time lag and amplitude of the autocorrelation
-    #----------------------------------------------------------
-    local_min=argrelmin(autoc)
-    index=(local_min[0][0])
-    r=(autoc[index])
-    r=math.sqrt(r**2)
-    tlag= time[index]
+    n = len(StackedData)
+    fmax = 1 / (2.0 * dt)
+    df = fmax / (n / 2)
+    f = np.hstack((df * np.arange(0, n//2), df * np.arange(-n//2 + 1, 0)))
+    nf = n // 2 + 1
+    dw = 2.0 * np.pi * df
+    w = dw * np.hstack((np.arange(0, n//2), np.arange(-n//2 + 1, 0)))
+    filtered_rf = np.zeros_like(StackedData)
+    D = StackedData
+    D = D - np.mean(D)
+    D = detrend(D)
+    #-----------------------------------------------
+    # Calculate the autocorrelation
+    ac = np.correlate(D, D, mode='full')
+    ac = ac / np.max(ac)
+    ac = ac[n-1:2*n-1]
+
+    rac = -ac
+    locs, _ = find_peaks(rac)
+    if len(locs) == 0:
+        r0 = 0
+        tlag = 0
+        print('No reverberation detected')
+    else:
+        r0 = np.abs(rac[locs[0]])
[…]
