### 2024-02-14 3ef2367
Updated SynthAnalysis Module
