### 2024-02-14 de41286
Updated the H and K arrays in test
