+++ src/rfsed/tests/test_WaveformFittingMultiproc.py
@@ -17,10 +17,8 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
         gauparameter=1.25
         savepath=save_tests()
         rayp=0.04
-        # KMoho= np.linspace(1.65,1.95,121)
-        # HMoho=np.linspace(20,60,201)
-        KMoho= np.linspace(1.65,1.95,5)
-        HMoho=np.linspace(20,60,5)
+        KMoho= np.linspace(1.65,1.95,121)
+        HMoho=np.linspace(20,60,201)
         wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
         WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gauparameter)
 
