+++ src/rfsed/tests/test_hkYu.py
@@ -13,17 +13,20 @@ except ImportError:
 class ReverbFilter(unittest.TestCase):
     def test_Resonance_Filt(self):
         rfstream = rfMoho_example()
-        self.assertTrue(Resonance_Filt(rfstream))
+        preonset=10
+        self.assertTrue(Resonance_Filt(rfstream, preonset))
 
     def test_hkYu(self):
         rfstream = rfMoho_example()
-        FilteredRF= Resonance_Filt(rfstream)
+        preonset=10
+        FilteredRF= Resonance_Filt(rfstream, preonset)
         hkYu(FltResult=FilteredRF, rayp=0.04, HSubSed=np.linspace(20,60,201), KSubSed=np.linspace(1.65,1.95,121), 
                HSed=np.linspace(0,10,201), KSed=np.linspace(1.65,2.25,201), VpMoho=6.9, VpSed= 2.5,  
                w1SubSed=0.6, w2SubSed=0.3, w3SubSed=0.1, w1Sed=0.6, w2Sed=0.3, w3Sed=0.1)  
     def test_plothkYu(self):
         rfstream = rfMoho_example()
-        FilteredRF= Resonance_Filt(rfstream)
+        preonset=10
+        FilteredRF= Resonance_Filt(rfstream, preonset)
         savepath=save_tests()
         HKResults=hkYu(FltResult=FilteredRF, rayp=0.04, HSubSed=np.linspace(20,60,201), KSubSed=np.linspace(1.65,1.95,121), 
                 HSed=np.linspace(0,10,201), KSed=np.linspace(1.65,2.25,201), VpMoho=6.9, VpSed= 2.5,  
