+++ src/rfsed/tests/test_hkZhu.py
@@ -12,24 +12,26 @@ except ImportError:
 class TesthkZhu(unittest.TestCase):
     def test_hk(self):
         rfstream = rfMoho_example()
+        preonset=10
         staname='OPLO'
         rfstreams = rfstream.select(component='R', station=staname)
         K= np.linspace(1.65,1.95,121)
         H=np.linspace(20,60,201)
         Vp=6.9
         w1, w2, w3 =[0.6, 0.3, 0.1]
-        self.assertTrue(hkZhu.hk(rfstreams, layer='Moho', stack=True, w1 = w1, w2 = w2, w3 = w3, K= K, H=H, Vp = Vp))
+        self.assertTrue(hkZhu.hk(rfstreams, preonset, layer='Moho', stack=True, w1 = w1, w2 = w2, w3 = w3, K= K, H=H, Vp = Vp))
     
     def test_plothk(self):
         rfstream = rfMoho_example()
         staname='OPLO'
+        preonset=10
         rfstreams = rfstream.select(component='R', station=staname)
         K= np.linspace(1.65,1.95,121)
         H=np.linspace(20,60,201)
         Vp=6.9
         w1, w2, w3 =[0.6, 0.3, 0.1]
         savepath=save_tests()
-        HKResult= hkZhu.hk(rfstreams, layer='Moho', stack=True, w1 = w1, w2 = w2, w3 = w3, K= K, H=H, Vp = Vp)
+        HKResult= hkZhu.hk(rfstreams, preonset, layer='Moho', stack=True, w1 = w1, w2 = w2, w3 = w3, K= K, H=H, Vp = Vp)
         hkZhu.plothk(HKResult=HKResult, g = [75.,10., 15., 2.5], rmneg = None,savepath=savepath, format = 'jpg')
 
 def suite():
