+++ src/rfsed/tests/test_hkSeqYeck.py
@@ -12,6 +12,7 @@ except ImportError:
 class hkSeqYeck(unittest.TestCase):
     def test_hkSeq(self):
         rfstSed = rfSed_example()
+        preonset=10
         staname='OPLO'
         rfstreamSed = rfstSed.select(component='R', station=staname)
         KSed= np.linspace(1.65,2.25,201)
@@ -24,12 +25,13 @@ class hkSeqYeck(unittest.TestCase):
         HMoho=np.linspace(20,60,201)
         VpMoho=6.9
         w1Moho, w2Moho, w3Moho = [0.6, 0.3, 0.1]
-        self.assertTrue(hkSeq(rfstreamSed, rfstreamMoho,  w1Sed = w1Sed, 
+        self.assertTrue(hkSeq(rfstreamSed, rfstreamMoho,  preonset, w1Sed = w1Sed, 
              w2Sed = w2Sed, w3Sed=w3Sed, KSed=KSed, HSed=HSed, VpSed=VpSed,
              w1Moho = w1Moho, w2Moho = w2Moho, w3Moho=w3Moho, KMoho=KMoho, 
              HMoho=HMoho, VpMoho=VpMoho, stack = False))
     def test_plotSeqhk(self):
         rfstSed = rfSed_example()
+        preonset=10
         staname='OPLO'
         rfstreamSed = rfstSed.select(component='R', station=staname)
         KSed= np.linspace(1.65,2.25,201)
@@ -43,7 +45,7 @@ class hkSeqYeck(unittest.TestCase):
         VpMoho=6.9
         w1Moho, w2Moho, w3Moho = [0.6, 0.3, 0.1]
         savepath=save_tests()
-        SequentialHKResult=hkSeq(rfstreamSed, rfstreamMoho,  w1Sed = w1Sed, 
+        SequentialHKResult=hkSeq(rfstreamSed, rfstreamMoho, preonset, w1Sed = w1Sed, 
              w2Sed = w2Sed, w3Sed=w3Sed, KSed=KSed, HSed=HSed, VpSed=VpSed,
              w1Moho = w1Moho, w2Moho = w2Moho, w3Moho=w3Moho, KMoho=KMoho, 
              HMoho=HMoho, VpMoho=VpMoho, stack = False)
