### 2024-02-14 0ce7df1
Changed 'gauparameter' variable naming to 'gaussian'
