+++ src/rfsed/tests/test_WaveformFittingMultiproc.py
@@ -14,27 +14,27 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
         VpSed=3.3
         Sedthick=1.8
         VpCrust = 6.9
-        gauparameter=1.25
+        gaussian=1.25
         savepath=save_tests()
         rayp=0.04
         KMoho= np.linspace(1.65,1.95,121)
         HMoho=np.linspace(20,60,201)
         wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-        WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gauparameter)
+        WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gaussian)
 
     def test_run_waveformfitting(self):
             rfstream= rfMoho_example()
             VpSed=3.3
             Sedthick=1.8
             VpCrust = 6.9
-            gauparameter=1.25
+            gaussian=1.25
             savepath=save_tests()
             rayp=0.04
             KMoho= np.linspace(1.65,1.95,121)
             HMoho=np.linspace(20,60,201)
             nproc=25
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-            ModelParams=WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gauparameter)
+            ModelParams=WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gaussian)
             run_waveformfitting(nproc, HMoho, ModelParams)
 
     def test_plotbestmodel(self):
@@ -42,7 +42,7 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             VpSed=3.3
             Sedthick=1.8
             VpCrust = 6.9
-            gauparameter=1.25
+            gaussian=1.25
             savepath=save_tests()
             rayp=0.04
             KMoho= np.linspace(1.65,1.95,121)
@@ -50,7 +50,7 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             nproc=25
             format='pdf'
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-            ModelParams=WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gauparameter)
+            ModelParams=WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gaussian)
             Results=run_waveformfitting(nproc, HMoho, ModelParams)
             plotbestmodel(Results, ModelParams, wtCorr, wtRMSE, wtPG, savepath, format)
         
