+++ src/rfsed/tests/test_ReverbFilter.py
@@ -12,11 +12,13 @@ except ImportError:
 class ReverbFilter(unittest.TestCase):
     def test_Resonance_Filt(self):
         rfstream = rfMoho_example()
-        self.assertTrue(Resonance_Filt(rfstream))
+        preonset=10
+        self.assertTrue(Resonance_Filt(rfstream, preonset))
     def test_plotfiltrf(self):
         rfstream = rfMoho_example()
+        preonset=10
         savepath=save_tests()
-        FilteredRF= Resonance_Filt(rfstream)
+        FilteredRF= Resonance_Filt(rfstream, preonset)
         plotfiltrf(FilteredRF, savepath, format = 'jpg')        
 
 def suite():
