+++ src/rfsed/tests/test_WaveformFittingMultiproc.py
@@ -20,7 +20,8 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
         KMoho= np.linspace(1.65,1.95,121)
         HMoho=np.linspace(20,60,201)
         wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-        WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gaussian)
+        WaveformPara(rfstream=rfstream, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
+                                     rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
 
     def test_run_waveformfitting(self):
             rfstream= rfMoho_example()
@@ -34,7 +35,8 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             HMoho=np.linspace(20,60,201)
             nproc=25
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-            ModelParams=WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gaussian)
+            ModelParams=WaveformPara(rfstream=rfstream, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
+                                     rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
             run_waveformfitting(nproc, HMoho, ModelParams)
 
     def test_plotbestmodel(self):
@@ -50,7 +52,8 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             nproc=25
             format='pdf'
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-            ModelParams=WaveformPara(rfstream, Sedthick, VpSed, VpCrust, rayp, KMoho, HMoho, gaussian)
+            ModelParams=WaveformPara(rfstream=rfstream, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
+                                     rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
             Results=run_waveformfitting(nproc, HMoho, ModelParams)
             plotbestmodel(Results, ModelParams, wtCorr, wtRMSE, wtPG, savepath, format)
         
