### 2024-02-14 dd63206
Add preonset to WaveformFitting Tests
