+++ src/rfsed/tests/test_WaveformFittingMultiproc.py
@@ -11,6 +11,7 @@ except ImportError:
 class TestWaveformFittingMultiproc(unittest.TestCase):
     def test_WaveformPara(self):
         rfstream= rfMoho_example()
+        preonset=10
         VpSed=3.3
         Sedthick=1.8
         VpCrust = 6.9
@@ -20,11 +21,12 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
         KMoho= np.linspace(1.65,1.95,121)
         HMoho=np.linspace(20,60,201)
         wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-        WaveformPara(rfstream=rfstream, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
+        WaveformPara(rfstream=rfstream, preonset=preonset, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
                                      rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
 
     def test_run_waveformfitting(self):
             rfstream= rfMoho_example()
+            preonset=10
             VpSed=3.3
             Sedthick=1.8
             VpCrust = 6.9
@@ -35,12 +37,13 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             HMoho=np.linspace(20,60,201)
             nproc=25
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-            ModelParams=WaveformPara(rfstream=rfstream, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
+            ModelParams=WaveformPara(rfstream=rfstream, preonset=preonset, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
                                      rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
             run_waveformfitting(nproc, HMoho, ModelParams)
 
     def test_plotbestmodel(self):
             rfstream= rfMoho_example()
+            preonset=10
             VpSed=3.3
             Sedthick=1.8
             VpCrust = 6.9
@@ -52,7 +55,7 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             nproc=25
             format='pdf'
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
-            ModelParams=WaveformPara(rfstream=rfstream, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
+            ModelParams=WaveformPara(rfstream=rfstream, preonset=preonset, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
                                      rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
             Results=run_waveformfitting(nproc, HMoho, ModelParams)
             plotbestmodel(Results, ModelParams, wtCorr, wtRMSE, wtPG, savepath, format)
