### 2024-02-14 c8829f5
Updated ExtractEq Module
