### 2024-02-14 47bde30
Updated ExtractEq Module
