### 2024-02-15 77fdae4
Updated ExtractEq Module
