### 2024-02-16 35c2ec5
Removed padding in resonance filter and use appropriate length in frequency domain
