+++ src/rfsed/ReverbFilter.py
@@ -64,10 +64,10 @@ def Resonance_Filt(rfstream, preonset):
     n = len(StackedData)
     fmax = 1 / (2.0 * dt)
     df = fmax / (n / 2)
-    f = np.hstack((df * np.arange(0, n//2), df * np.arange(-n//2 + 1, 0)))
+    f = np.hstack((df * np.arange(0, n//2 +1), df * np.arange(-n//2 + 1, 0)))
     nf = n // 2 + 1
     dw = 2.0 * np.pi * df
-    w = dw * np.hstack((np.arange(0, n//2), np.arange(-n//2 + 1, 0)))
+    w = dw * np.hstack((np.arange(0, n//2+1), np.arange(-n//2 + 1, 0)))
     filtered_rf = np.zeros_like(StackedData)
     D = StackedData
     D = D - np.mean(D)
@@ -89,7 +89,6 @@ def Resonance_Filt(rfstream, preonset):
         tlag = locs[0] * dt
 
     resonanceflt = (1 + r0 * np.exp(-1j * w * tlag))
-    resonanceflt = np.hstack((resonanceflt, np.zeros(n - len(resonanceflt))))  # Pad filter to match length of FFT result
     filtered_rf = np.real(ifft(fft(D) * resonanceflt))
     # filtered_rf = filtered_rf / np.max(filtered_rf)
     #------------------------------------------------------------------------------
