### 2024-02-16 abbc0b5
Removed padding in resonance filter and use appropriate length in frequency domain
