### 2024-02-16 dc3adc3
Plot only real part of resonance filter
