+++ src/rfsed/SynthAnalysis.py
@@ -631,6 +631,7 @@ def plotfiltSynthrf(FilteredRF, savepath, format = 'jpg'):
     rf=FilteredRF['rf']
     filtered_rf=FilteredRF['filteredrf']
     resonanceflt=FilteredRF['resonancefilter']
+    resonanceflt=np.real(resonanceflt)
     time=FilteredRF['time']
     autoc=FilteredRF['autoc']
     suff='SyntheticRF'
