### 2024-02-16 1c5b962
Editted Waveform Fitting Module
