+++ src/rfsed/SynthAnalysis.py
@@ -14,8 +14,6 @@ from rfsed.SynRF.FwdRF import SynRF
 import matplotlib.gridspec as gridspec
 import matplotlib.transforms as mtransforms
 from matplotlib.legend_handler import HandlerLine2D as HL
-import seaborn
-seaborn.set_style("darkgrid", {"axes.facecolor": ".85"})
 from scipy import signal
 from scipy.signal import argrelextrema, argrelmin, argrelmax
 import math
