+++ src/rfsed/tests/test_ReverbFilter.py
@@ -1,7 +1,7 @@
 import unittest
 import warnings
 from rfsed import ReverbFilter
-from rfsed.ReverbFilter import Resonance_Filt, plotfiltrf
+from rfsed.ReverbFilter import ResonanceFilt, plotfiltrf
 import numpy as np
 from rfsed.util import rfMoho_example, save_plot, save_tests
 try:
@@ -13,12 +13,12 @@ class ReverbFilter(unittest.TestCase):
     def test_Resonance_Filt(self):
         rfstream = rfMoho_example()
         preonset=10
-        self.assertTrue(Resonance_Filt(rfstream, preonset))
+        self.assertTrue(ResonanceFilt(rfstream, preonset))
     def test_plotfiltrf(self):
         rfstream = rfMoho_example()
         preonset=10
         savepath=save_tests()
-        FilteredRF= Resonance_Filt(rfstream, preonset)
+        FilteredRF= ResonanceFilt(rfstream, preonset)
         plotfiltrf(FilteredRF, savepath, format = 'jpg')        
 
 def suite():
