+++ src/rfsed/hkYu.py
@@ -223,8 +223,8 @@ def plothkYu(hkYuResult, savepath, g = [75.,10., 15., 2.5], rmneg = None, format
     tlag=hkYuResult['tlag']
     VpMoho=hkYuResult['VpMoho']
     VpSed=hkYuResult['VpSed']
-    HSed=hkYuResult['HSed']
-    KSed=hkYuResult['KSed']
+    # HSed=hkYuResult['HSed']
+    # KSed=hkYuResult['KSed']
     HMoho=hkYuResult['HMoho']
     KMoho=hkYuResult['KMoho']
     rfdata=hkYuResult['FltRF']
