+++ src/rfsed/tests/test_hkYu.py
@@ -1,7 +1,7 @@
 import unittest
 import warnings
 from rfsed import ReverbFilter
-from rfsed.ReverbFilter import Resonance_Filt, plotfiltrf
+from rfsed.ReverbFilter import ResonanceFilt, plotfiltrf
 from rfsed.hkYu import hkYu, plothkYu
 import numpy as np
 from rfsed.util import rfMoho_example, save_plot, save_tests
@@ -14,12 +14,12 @@ class ReverbFilter(unittest.TestCase):
     def test_Resonance_Filt(self):
         rfstream = rfMoho_example()
         preonset=10
-        self.assertTrue(Resonance_Filt(rfstream, preonset))
+        self.assertTrue(ResonanceFilt(rfstream, preonset))
 
     def test_hkYu(self):
         rfstream = rfMoho_example()
         preonset=10
-        FilteredRF= Resonance_Filt(rfstream, preonset)
+        FilteredRF= ResonanceFilt(rfstream, preonset)
         hkYu(FltResult=FilteredRF, rayp=0.04, HSubSed=np.linspace(20,60,201), KSubSed=np.linspace(1.65,1.95,121), 
              VpMoho=6.9, VpSed= 2.5, VsSed=1.4, w1SubSed=0.6, w2SubSed=0.3, w3SubSed=0.1)  
         # hkYu(FltResult=FilteredRF, rayp=0.04, HSubSed=np.linspace(20,60,201), KSubSed=np.linspace(1.65,1.95,121), 
@@ -28,7 +28,7 @@ class ReverbFilter(unittest.TestCase):
     def test_plothkYu(self):
         rfstream = rfMoho_example()
         preonset=10
-        FilteredRF= Resonance_Filt(rfstream, preonset)
+        FilteredRF= ResonanceFilt(rfstream, preonset)
         savepath=save_tests()
         HKResults=hkYu(FltResult=FilteredRF, rayp=0.04, HSubSed=np.linspace(20,60,201), KSubSed=np.linspace(1.65,1.95,121), 
                 HSed=np.linspace(0,10,201), KSed=np.linspace(1.65,2.25,201), VpMoho=6.9, VpSed= 2.5,  
