### 2024-03-17 #6300 @elbeejay
Hello @trichter and @paudetseis, I'm reaching out to ask if you would be able to review the JOSS submission "rfsed: Receiver function analysis and dealing with sediment effects"

I am aware you both have experience publishing your work in JOSS, so I believe you'll be familiar with the process. I believe this submission aligns well with your subject-matter expertise, and would appreciate your reviews if you have the time.

Please let me know and reach out if you have any questions. Thanks!
