### 2024-03-24 #6300 @elbeejay
@rbeucher, I'm reaching out to ask if you'd be interested in reviewing this submission for JOSS titled "rfsed: Receiver function analysis and dealing with sediment effects". From what I can tell you have experience in the field of geodynamics and are comfortable programming in Python. I know you've also taken part in the JOSS review process as both a reviewer and as an author. Please take a look at this article and let us know if you are able and willing to serve as a peer-reviewer.

Thanks again for taking a moment to consider this, feel free to contact me if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know! If interested we can figure out how to set up a "co-review" for a colleague of yours that is a more junior or inexperienced member of the community.

Let me know, thanks.
Jay
