### 2024-03-28 #6300 @elbeejay
@sannecottaar and @bobmyhill I'm reaching out to ask if you'd be interested in reviewing the submission "rfsed: Receiver function analysis and dealing with sediment effects" for JOSS. I expect you're both familiar with JOSS and our publication process from your recent publication on the `BurnMan` toolkit. Please take a look at this submission and let us know here if you'd be able to provide a review, thanks!
