### 2024-03-28 #6300 @bobmyhill
I don't think I count as an expert in the field, so I shouldn't review by JOSS guidelines. But here are a couple of things that I noticed:

- rf is published on PyPI - is there a reason you copy the code here? It is almost always better to add existing modules as dependencies.
- SeisPy is also published on PyPI. If you've modified the code because it needs correcting or improving, why not submit a PR to the SeisPy team? SeisPy is actively maintained.
- The environment.yml file is very restrictive. If everyone required exact versions of every dependency as done here, it would be very hard to set up multi-module python environments.
- The structure of the repo is unusual - tests and examples aren't normally stored in the src directory. Typically, module structure looks something like
project/
├ package/
├ ├ __init__.py
├ ├ code1.py
├ ├ code2.py
├ examples/
├ tests/
├ ├ __init__.py
├ ├ test1.py
├ setup.py
