### 2024-04-12 #6612 @elbeejay
Hi @akinremisa, @paudetseis, and @alistairboyce11,

Thank you all for participating in the JOSS review process. At this time, I will be asking our generous reviewers to begin their reviews.

@paudetseis and @alistairboyce11, the instructions for making your checklist and conducting your reviews should be in the first comment of this issue, but please do not hesitate to reach out to me and I can definitely help you navigate this process. We prefer that larger comments be made as issues in the rfsed repository itself, and that this issue thread focus more on the JOSS process. This is to hopefully keep issues related to the content and code tied directly to the code repository itself. If possible, please link to this JOSS review issue in any rfsed repository issue you create, so that we have some connectivity there.

The JOSS review process can be 'conversational' in a way that is atypical for peer-reviewed journals. By this, we mean that you as the reviewer can have a direct dialog with the author as they respond to and make edits due to a reviewer comment you have made. We ask that you try to take a first pass and make comments within 6 weeks. If that is not possible it is not a problem, we just ask that you all try to keep the lines of communication open and communicate in this review thread if you're likely to miss the target deadline.

I will be asking the bot to send reminders in a few weeks to help keep this review from falling off of anyone's to-do list.

Thanks again!
Jay
