### 2024-04-12 akinremisa/rfsed#1 @elbeejay
JOSS Pre-Review Comments
Hi @akinremisa - during the JOSS pre-review phase we got a few good comments from @bobmyhill that I didn't want us to skip-over during the review phase. I am pasting them below. If you could address them in this issue and/or make any edits necessary to your package that'd be great.

> - rf is published on PyPI - is there a reason you copy the code here? It is almost always better to add existing modules as dependencies.
> - SeisPy is also published on PyPI. If you've modified the code because it needs correcting or improving, why not submit a PR to the SeisPy team? SeisPy is actively maintained.
> - The environment.yml file is very restrictive. If everyone required exact versions of every dependency as done here, it would be very hard to set up multi-module python environments.
> - The structure of the repo is unusual - tests and examples aren't normally stored in the src directory. Typically, module structure looks something like
> project/
> ├ package/
> ├ ├ init.py
> ├ ├ code1.py
> ├ ├ code2.py
> ├ examples/
> ├ tests/
> ├ ├ init.py
> ├ ├ test1.py
> ├ setup.py

This issue relates to https://github.com/openjournals/joss-reviews/issues/6612
