### 2024-04-20 a9fe87d
changed paths to data and directory to save outputs in the util.py file
