### 2024-04-20 7574196
Editted the environment.yml file to be require exact versions of dependencies
