### 2024-04-20 e20a333
Editted the environment.yml file to be require exact versions of dependencies
