+++ environment.yml
@@ -45,7 +45,7 @@ dependencies:
     - lxml>=4.9.4
     - markupsafe>=2.1.3
     - matplotlib>=3.8.2
-    - numpy>=1.26.2
+    - numpy<=1.24.9
     - obspyh5>=0.6.0
     - owslib>=0.29.3
     - packaging>=23.2
