### 2024-04-22 fbf49e5
Modify FwdRF to synrf and implement the forward calculator to account for any number of layers in the synthetic model
