### 2024-04-22 bcdc205
change the implementation of FwdRF to synrf in the  modules
