### 2024-04-22 5492c9c
Modified the manuscript and bib files
