+++ joss/paper.bib
@@ -32,20 +32,6 @@
    volume = {103},
    year = {2013},
 }
-@article{Xu2022,
-    author = {Xu, Mijian and He, Jing},
-    title = "{Seispy: Python Module for Batch Calculation and Postprocessing of Receiver Functions}",
-    journal = {Seismological Research Letters},
-    volume = {94},
-    number = {2A},
-    pages = {935-943},
-    year = {2022},
-    month = {12},
-    issn = {0895-0695},
-    doi = {10.1785/0220220288},
-    url = {https://doi.org/10.1785/0220220288},
-    eprint = {https://pubs.geoscienceworld.org/ssa/srl/article-pdf/94/2A/935/5793459/srl-2022288.1.pdf},
-}
 @article{Yu2015,
    author = {Youqiang Yu and Jianguo Song and Kelly H. Liu and Stephen S. Gao},
    doi = {10.1002/2014JB011610},
