### 2024-04-26 akinremisa/rfsed#1 @akinremisa
•	rf is published on PyPI - is there a reason you copy the code here? It is almost always better to add existing modules as dependencies.
_rf is published under The MIT License (MIT) Copyright (c) 2013-2019 Tom Eulenfeld with permission that includes use, copy, modify, merge, publish rights. We have included a modified version of rf in this repository, especially in the class dealing with the receiver function to calculate extra parameters that are necessary for rfsed processing but not directly relevant to the rf package itself. Does this not align with JOSS publication requirements? If not, we can include rf as a dependency._
•	SeisPy is also published on PyPI. If you've modified the code because it needs correcting or improving, why not submit a PR to the SeisPy team? SeisPy is actively maintained.
_This is now solved. We now only use modified functions (from seispy- under GNU GENERAL PUBLIC LICENSE) for calculating synthetic receiver function, in such a way that it directly fits the rfsed implementations. This is properly referenced in the repository._
•	The environment.yml file is very restrictive. If everyone required exact versions of every dependency as done here, it would be very hard to set up multi-module python environments.
_The environment.yml file is now edited to not require exact versions of the dependencies of the package. We tested it to ensure that the rfsed installation still works perfectly._
•	The structure of the repo is unusual - tests and examples aren't normally stored in the src directory.
_The structure of the package directory is now re-ordered such that the data, tests, and examples directories are now in the same level as the src directory._
