### 2024-04-26 akinremisa/rfsed#1 @bobmyhill
@akinremisa, thanks for your efforts in modifying the package!

Regarding `rf`, your use is not against JOSS requirements, but it does place significant burden of responsibility on you to fix bugs and keep the copied package functional as python evolves. Your changes to the package appear to be minimal, and I don't see anywhere where you couldn't import the required functions from `rf`.

A related comment regarding both `rf` and  `SeisPy` - there's nothing inherently wrong with what you've done, but for open-source academic software developers, others using our software is (a) a major motivator for continued development and (b) a key indicator of esteem (which helps with grant applications, promotions etc.). If everyone copies our code directly into their own repositories rather than adding it as a dependency, we don't get full credit for our work.
