+++ src/rfsed/rf/imaging.py
@@ -1,379 +1,0 @@
-# Copyright 2013-2019 Tom Eulenfeld, MIT license
-"""
-All objects and functions below are after the Tom Eulenfeld's
-rf project License below.
-
-The MIT License (MIT)
-
-Copyright (c) 2013-2019 Tom Eulenfeld
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
-of the Software, and to permit persons to whom the Software is furnished to do
-so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS
-FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-"""
-"""
-Functions for receiver function plotting.
-"""
-import warnings
-
-import matplotlib.patheffects as PathEffects
-from matplotlib.ticker import (AutoMinorLocator, FixedLocator, FixedFormatter,
-                               MaxNLocator)
-import matplotlib.pyplot as plt
-import numpy as np
-
-
-def _label(stream):
-    label_fmts = ['{network}.{station}.{location}.{channel}',
-                  '{network}.{station}.{location}.{channel:.2}?',
-                  '{network}.{station}']
-    for label_fmt in label_fmts:
-        labelset = {label_fmt.format(**tr.stats) for tr in stream}
-        if len(labelset) == 1:
-            return labelset.pop()
-    return ''
-
-
-def plot_rf(stream, fname=None, fig_width=7., trace_height=0.5,
-            stack_height=0.5, dpi=None,
-            scale=1, fillcolors=(None, None), trim=None,
-            info=(('back_azimuth', u'baz (°)', 'C0'),
-                  ('distance', u'dist (°)', 'C3')),
-            show_vlines=False):
-    """
-    Plot receiver functions.
-
-    :param stream: stream to plot
-    :param fname: filename to save plot to. Can be None. In this case
-        the figure is left open.
-    :param fig_width: width of figure in inches
-    :param trace_height: height of one trace in inches
-    :param stack_height: height of stack axes in inches
-    :param dpi: dots per inch for the created figure
-    :param scale: scale for individual traces
-    :param fillcolors: fill colors for positive and negative wiggles
-    :param trim: trim stream relative to onset before plotting using
-         `~.rfstream.RFStream.slice2()`
-    :param info: Plot one additional axes showing maximal two entries of
-        the stats object. Each entry in this list is a list consisting of
-        three entries: key, label and color.
-        info can be None. In this case no additional axes is plotted.
-    :param show_vlines: If True, show vertical alignment grid lines on plot
[…]
