+++ src/rfsed/rf/profile.py
@@ -1,155 +1,0 @@
-# Copyright 2013-2019 Tom Eulenfeld, MIT license
-"""
-All objects and functions below are after the Tom Eulenfeld's
-rf project License below.
-
-The MIT License (MIT)
-
-Copyright (c) 2013-2019 Tom Eulenfeld
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
-of the Software, and to permit persons to whom the Software is furnished to do
-so, subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS
-FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-"""
-"""
-Functions for receiver function profile calculation.
-"""
-import numpy as np
-from rfsed.rf.util import _add_processing_info, direct_geodetic
-
-
-_LARGE_BOX_WIDTH = 2000
-
-
-def _get_box(latlon0, azimuth, length, width=_LARGE_BOX_WIDTH, offset=0):
-    """Create a single box."""
-    from shapely.geometry import Polygon
-    start = direct_geodetic(latlon0, azimuth, offset)
-    azis = ((azimuth - 90) % 360, azimuth,
-            (azimuth + 90) % 360, (azimuth + 180) % 360)
-    dists = (width/2, length, width, length)
-    latlon = start
-    corners = []
-    for a, d in zip(azis, dists):
-        latlon = direct_geodetic(latlon, a, d)
-        corners.append(latlon[::-1])
-    box = {'poly': Polygon(corners),
-           'length': length,
-           'pos': offset + length/2,
-           'latlon': direct_geodetic(start, azimuth, length/2)}
-    return box
-
-
-def get_profile_boxes(latlon0, azimuth, bins, width=_LARGE_BOX_WIDTH):
-    """
-    Create 2D boxes for usage in `profile()` function.
-
-    :param tuple latlon0: coordinates of starting point of profile
-    :param azimuth: azimuth of profile direction
-    :param tuple bins: Edges of the distance bins in km (e.g. (0, 10, 20, 30))
-    :param width: width of the boxes in km (default: large)
-    :return: List of box dicts. Each box has the entries
-        'poly' (shapely polygon with lonlat corners), 'length' (length in km),
-        'pos' (midpoint of box in km from starting coordinates),
-        'latlon' (midpoint of box as coordinates)
-    """
-    boxes = []
-    for i in range(len(bins)-1):
-        length = bins[i+1] - bins[i]
-        box = _get_box(latlon0, azimuth, length, width, offset=bins[i])
-        if i == 0:
-            box['profile'] = {}
-            box['profile']['latlon'] = latlon0
-            box['profile']['azimuth'] = azimuth
[…]

[…]
