+++ src/rfsed/WaveformFitting.py
@@ -93,7 +93,7 @@ def WaveformFitting(rfstream, preonset, HSed, VpSed, VpCrust, rayp, KMoho, HMoho
             vp = np.array([VpSed, VpCrust, 8.045])
             vs = np.array([VsSed, (VpCrust/Ktemp), 4.49])
             rho = np.array([SedDen, CrustDen, 3.299])
-            Synth=SynRF(depth, vp, vs, rho, rayp, dt=delta, npts=n, ipha=1)
+            Synth=synrf(depth, vp, vs, rho, rayp, dt=delta, npts=n, ipha=1)
             Synth.run_fwd()
             Synth.filter(freqmin=0.05, freqmax=1.25, order=2, zerophase=True)
             rf_synth=Synth.run_deconvolution(pre_filt=[0.05, 1.25], shift=delay, gaussian=gaussian)
@@ -275,7 +275,7 @@ def PlotWaveformFitting(WaveformFitting, wtCorr, wtRMSE, wtPG, savepath, format)
         vp = np.array([VpSed, VpCrust, 8.045])
         vs = np.array([VsSed, (VpCrust/K), 4.49])
         rho = np.array([SedDen, CrustDen, 3.299])
-        Synth=SynRF(depth, vp, vs, rho, rayp, dt=delta, npts=n, ipha=1)
+        Synth=synrf(depth, vp, vs, rho, rayp, dt=delta, npts=n, ipha=1)
         Synth.run_fwd()
         Synth.filter(freqmin=0.05, freqmax=1.25, order=2, zerophase=True)
         rf_synth=Synth.run_deconvolution(pre_filt=[0.05, 1.25], shift=delay, gaussian=gaussian)
