+++ setup.cfg
@@ -19,6 +19,7 @@ package_dir =
 	= src
 python_requires = >=3.6, <3.11
 install_requires = 
+	rf>=1.0.3
 	decorator
 	matplotlib
 	numpy
