+++ environment.yml
@@ -59,6 +59,7 @@ dependencies:
     - pytz>=2023.3.post1
     - pyyaml>=6.0.1
     - requests>=2.31.0
+    - rf>=1.0.3
     - scikit-learn>=1.3.2
     - seaborn>=0.13.0
     - six>=1.16.0
