+++ setup.cfg
@@ -42,6 +42,7 @@ install_requires =
 	colorama>=0.1.6
 	plotly>=5.17.0
 	psutil>=5.9.6
+	ipykernel>=6.29.4
 tests_require = 
 	pytest
 	py
