+++ environment.yml
@@ -40,6 +40,7 @@ dependencies:
     - imagesize>=1.4.1
     - importlib-metadata>=7.0.1
     - importlib-resources>=6.1.1
+    - ipykernel>=6.29.4
     - jinja2>=3.1.2
     - kiwisolver>=1.4.5
     - lxml>=4.9.4
