+++ joss/paper.bib
@@ -1,12 +1,11 @@
 @article{Akinremi2024,
    author = {Stephen Akinremi and Mark van der Meijde and Christine Thomas and Juan C. Afonso and Elmer Ruigrok and Islam Fadel},
-   doi = {under review},
-   issn = {under review},
-   issue = {under review},
+   doi = {10.1029/2023JB028393},
+   issue = {5},
    journal = {Journal of Geophysical Research: Solid Earth},
-   pages = {under review},
+   pages = {1-11},
    title = {Waveform fitting of receiver functions for enhanced retrieval of crustal structure in the presence of sediments},
-   volume = {under review},
+   volume = {129},
    year = {2024},
 }
 @article{Eulenfeld2020,
