### 2024-05-08 akinremisa/rfsed#2 @paudetseis
JOSS Review Comment #1
Hi @akinremisa,

Thank you for making your code public. It looks like an interesting and useful application, and I would likely end up using it myself or recommending it to collaborators. However, I found the package to be deficient in several regards, and I strongly encourage you to address the following comments and suggestions:

- [x] Installing `rfsed` using `pip` works fine, but it isn't clear what to do once the package is installed this way. One can then import it using `import rfsed`, but then what? Because the software is not clearly developed as an API (although it *could* be, see below) but rather a collection of example notebooks that import the `rfsed` modules and run the functions, installing with `pip` makes little sense to me because those scripts are not available by importing `rfsed`, and the user does not have a clear idea where the `examples` folder is located (on my computer they are in `/opt/anaconda3/envs/rfsedtest/lib/python3.10/site-packages/rfsed/examples`). If you wish to make `rfsed` available via PyPI, you should have a function that installs the notebooks locally (e.g., ```import rfsed; rfsed.install_notebooks()```), and there should be documentation to explain this and how to run the tests, etc. Alternatively, you could remove `rfsed` from PyPI and only provide installation instructions through cloning.
- [x] Installing the package by cloning the repository and mirroring the `conda` environment did not work for me. It appears like the `environment.yml` file is too restrictive with specific package versions, and this install is not flexible enough. Most users will give up trying to fix the package versions, or you will end up spending lots of time trying to debug other people's environments. I recommend specifying which external packages are minimally required and installing them when creating the conda environment, e.g.,  `conda create -n rfsed python=3.9 obspyh5=0.6 ...`.
- [x] The Documentation link appears to link back to the README file and shows identical content. Proper documentation should describe how to use the code and not simply restate the README and point to the `examples` folder. In particular, this doesn't work when installing using PyPI. Please either make a full documentation or explain how to install and run the notebooks. However, even the notebooks should be edited to improve clarity (the code is difficult to read and contains a single comment per code cell). Writing proper documentation is essential; it improves the user experience and prevents you from receiving countless requests for help.
- [x] The `src/rfsed` folder contains the modules developed (each module being a collection of functions), but these are not listed anywhere, and the user doesn't know what they do. They appear to contain minimal documentation that can be consulted when typing, e.g., `from rfsed import ExtractEq; help(ExtractEq); help(ExtractEq.ExtractEq)`. Because the API information is available, it should be provided on the Documentation page without having to type `help()` on the command line. I recommend expanding the docstring to ensure the user knows exactly what the function does and how to call it properly. You could even add basic usage or specific examples directly in the docstring.
- [x] I highly recommend linking the documentation in a `docs` folder and building it using tools like Sphinx (https://www.sphinx-doc.org/en/master/) or Read the Docs (https://readthedocs.com). *Note*: I see that `sphinx` is required in the `environment.yml` file, but it's unclear how it's been used here.
- [x] I recommend improving the readability of your code by adhering to the `pep8` style guide (https://peps.python.org/pep-0008/). You could consider using `Black` for strict Python code formatting (https://github.com/psf/black)
- [x] Finally, you could check out other similar repositories to guide the restructuring of your code and address some of these comments. I can point you to my own repository `Telewavesim`, which is also a collection of modules that contain functions with example notebooks. This is by no means the only or best example, but it will give you some guidance for a package published in JOSS.

This issue relates to https://github.com/openjournals/joss-reviews/issues/6612
