### 2024-05-16 akinremisa/rfsed#3 @alistairboyce11
JOSS Review Comment #3
Review of 10.21105.joss.06612 rfsed: Receiver function analysis and dealing with sediment effects by Akinremi et al.,

Hi @akinremisa,

Thank you for your open source submission to JOSS. This is my first review for JOSS so please excuse my inexperience with the process. While I am not a user of RFs for crustal studies (more the mantle transition zone) I can see this kind of reverberation management providing useful tools for others. Overall, I would echo the vast majority of the points made by @paudetseis on [JOSS Review Comment #2](https://github.com/akinremisa/rfsed/issues/2), however I have tried to add a few additional points from my experiences. The companion paper is concisely written and serves the purpose stated by JOSS themselves.

* Firstly I ran into the same issues as @paudetseis when I tried to clone the directory directly from git. Please see his comment over on [JOSS Review Comment #2](https://github.com/akinremisa/rfsed/issues/2). I was able to install the rfsed package via pip, however.

* Testing: Its good to see the project includes the automated use of pytest. However, on running `pytest` on the codes in `rf/tests/*py`, the process proceeds rapidly but the following three tests (`test_WaveformFitting.py; test_ExtractEqMultiproc.py; test_WaveformFittingMultiproc.py`) in the `/tests/` appear to hang, or at least give the impression they are extremely slow to execute. Can this be modified so they execute on a more appropriate timescale?

* When I exclude the above three tests from the tests directory, pytest can complete its execution but it gives me the following output (which could be checked, after making a hacky installation below to execute the example notebooks, all tests passed, excluding those removed above):

======================================================= short test summary info ========================================================
FAILED rf/tests/test_batch.py::BatchTestCase::test_batch_command_interface_H5 - AttributeError: 'GrouperView' object has no attribute 'join'
FAILED rf/tests/test_batch.py::BatchTestCase::test_batch_command_interface_Q - AttributeError: 'GrouperView' object has no attribute 'join'
FAILED rf/tests/test_batch.py::BatchTestCase::test_batch_command_interface_SAC - AttributeError: 'GrouperView' object has no attribute 'join'
FAILED rf/tests/test_batch.py::BatchTestCase::test_entry_point - ImportError: Entry point ('console_scripts', 'rf') not found
FAILED rf/tests/test_profile.py::ProfileTestCase::test_profile - AttributeError: 'GrouperView' object has no attribute 'join'
ERROR rf/tests/test_batch.py::test_format
ERROR rf/tests/test_deconvolve.py::test_deconvolve_Lpeak
ERROR rf/tests/test_deconvolve.py::test_deconvolve_Qpeak
ERROR rf/tests/test_profile.py::test_io_header
ERROR rf/tests/test_rfstream.py::test_io_header
======================================== 5 failed, 34 passed, 6833 warnings, 5 errors in 32.63s ========================================

* In order to try to run the example notebooks I cloned the package from git and then heavily edited the environment.yml file (it was not clear how to access the examples using the pip install). I then removed all the version precision on the conda dependencies. Then commented the following six packages below so that the conda dependences could be installed.

  - _libgcc_mutex
  - _openmp_mutex
  - ld_impl_linux-64
  - libgcc-ng
  - libgomp
  - libstdcxx-ng

The component of the installation handled by pip hanged for a long period, even after trying to reverse the sign on the numpy package version or remove all package precisions entirely.

Instead, I split the install manually into a conda component and pip component (forcing a python version 3.9). Then I was able to run the example notebooks using the rfsed kernel I had created in VScode. Pleasingly, I was then able to run nearly all the notebooks in entirety without any apparent errors. An overall comment would be the notebooks appear a bit light in terms of comments to fully understand what each step is doing. Notebook 11 took over 8 minutes to produce an output while notebook 10 did not produce any output after 20 minutes. Perhaps the execution time can be reduced with one or two simpler examples.

Do let me know if something I have mentioned is not clear.

Best wishes,

Alistair

This issue relates to [openjournals/joss-reviews#6612](https://github.com/openjournals/joss-reviews/issues/6612)
