### 2024-05-20 #6612 @elbeejay
Thanks to both @paudetseis and @alistairboyce11 for opening up issues in the `rfsed` repository. @akinremisa you are welcome to provide responses to each in the respective issues and make changes to your repository as necessary. If you could comment here with some estimate of when you'll be able to read over and address the review comments I think that'd help both myself and our reviewers keep this on our schedules.

Thanks all!
