+++ setup.cfg
@@ -19,30 +19,18 @@ package_dir =
 	= src
 python_requires = >=3.6, <3.11
 install_requires = 
+	obspy>=1.0.3
 	rf>=1.0.3
-	decorator
 	matplotlib
 	numpy
 	scipy
-	obspy>=1.0.3
-	tqdm
+	colorama
 	OWSLib
-	cartopy
-	geographiclib
-	shapely
-	seaborn
-	scipy
-	scikit-learn
-	h5
-	h5py
-	obspyh5
-	tqdm
-	alabaster
-	sphinx
-	colorama>=0.1.6
-	plotly>=5.17.0
-	psutil>=5.9.6
-	ipykernel>=6.29.4
+	plotly
+	psutil
+	pyyaml
+	pytz>=2020.1
+	importlib-metadata
 tests_require = 
 	pytest
 	py
