### 2024-05-24 4229644
Remove environment.yml file
