+++ environment.yml
@@ -1,82 +1,0 @@
-name: rfsed
-channels:
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=5.1=1_gnu
-  - ca-certificates=2023.12.12=h06a4308_0
-  - ld_impl_linux-64=2.38=h1181459_1
-  - libffi=3.3=he6710b0_2
-  - libgcc-ng=11.2.0=h1234567_1
-  - libgomp=11.2.0=h1234567_1
-  - libstdcxx-ng=11.2.0=h1234567_1
-  - ncurses=6.4=h6a678d5_0
-  - openssl=1.1.1w=h7f8727e_0
-  - pip=23.3.1=py39h06a4308_0
-  - python=3.9.0=hdb3f193_2
-  - readline=8.2=h5eee18b_0
-  - setuptools=68.2.2=py39h06a4308_0
-  - sqlite=3.41.2=h5eee18b_0
-  - tk=8.6.12=h1ccaba5_0
-  - tzdata=2023c=h04d1e81_0
-  - wheel=0.41.2=py39h06a4308_0
-  - xz=5.4.5=h5eee18b_0
-  - zlib=1.2.13=h5eee18b_0
-  - pip:
-    - alabaster>=0.7.13
-    - babel>=2.14.0
-    - certifi>=2023.11.17
-    - charset-normalizer>=3.3.2
-    - colorama>=0.4.6
-    - contourpy>=1.2.0
-    - cycler>=0.12.1
-    - decorator>=5.1.1
-    - docutils>=0.20.1
-    - fonttools>=4.47.0
-    - geographiclib>=2.0
-    - greenlet>=3.0.3
-    - h5>=0.8.10
-    - idna>=3.6
-    - imagesize>=1.4.1
-    - importlib-metadata>=7.0.1
-    - importlib-resources>=6.1.1
-    - ipykernel>=6.29.4
-    - jinja2>=3.1.2
-    - kiwisolver>=1.4.5
-    - lxml>=4.9.4
-    - markupsafe>=2.1.3
-    - matplotlib>=3.8.2
-    - numpy<=1.24.9
-    - obspyh5>=0.6.0
-    - owslib>=0.29.3
-    - packaging>=23.2
-    - plotly>=5.18.0
-    - psutil>=5.9.7
-    - pygments>=2.17.2
-    - pyparsing>=3.1.1
-    - pyproj>=3.6.1
-    - pyshp>=2.3.1
-    - pytest>=7.4.4
-    - pytz>=2023.3.post1
-    - pyyaml>=6.0.1
-    - requests>=2.31.0
-    - rf>=1.0.3
-    - scikit-learn>=1.3.2
-    - seaborn>=0.13.0
-    - six>=1.16.0
-    - snowballstemmer>=2.2.0
-    - sphinx>=7.2.6
-    - sphinxcontrib-applehelp>=1.0.7
-    - sphinxcontrib-devhelp>=1.0.5
-    - sphinxcontrib-htmlhelp>=2.0.4
-    - sphinxcontrib-jsmath>=1.0.1
-    - sphinxcontrib-qthelp>=1.0.6
-    - sphinxcontrib-serializinghtml>=1.1.9
-    - sqlalchemy>=2.0.24
-    - tenacity>=8.2.3
-    - threadpoolctl>=3.2.0
-    - typing-extensions>=4.9.0
-    - urllib3>=2.1.0
-    - zipp>=3.17.0
-    - -e .
-
