### 2024-05-24 e501b06
Modify package dependencies
