+++ setup.cfg
@@ -55,6 +55,5 @@ tests =
 where = src
 
 [egg_info]
-tag_build = 
 tag_date = 0
 
