+++ setup.cfg
@@ -54,6 +54,4 @@ tests =
 [options.packages.find]
 where = src
 
-[egg_info]
-tag_date = 0
 
