### 2024-05-24 77cfd9b
Use smaller grid search discretization so the test runs faster
