+++ tests/test_WaveformFittingMultiproc.py
@@ -18,8 +18,11 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
         gaussian=1.25
         savepath=save_tests()
         rayp=0.04
-        KMoho= np.linspace(1.65,1.95,121)
-        HMoho=np.linspace(20,60,201)
+        # KMoho= np.linspace(1.65,1.95,121)
+        # HMoho=np.linspace(20,60,201)
+        # Use smaller grid search discretization so the test runs faster
+        KMoho= np.linspace(1.65,1.95,10)
+        HMoho=np.linspace(25,50,10)
         wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
         WaveformPara(rfstream=rfstream, preonset=preonset, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
                                      rayp=rayp, KMoho=KMoho, HMoho=HMoho, gaussian=gaussian)
@@ -33,8 +36,11 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             gaussian=1.25
             savepath=save_tests()
             rayp=0.04
-            KMoho= np.linspace(1.65,1.95,121)
-            HMoho=np.linspace(20,60,201)
+            # KMoho= np.linspace(1.65,1.95,121)
+            # HMoho=np.linspace(20,60,201)
+            # Use smaller grid search discretization so the test runs faster
+            KMoho= np.linspace(1.65,1.95,10)
+            HMoho=np.linspace(25,50,10)
             nproc=25
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
             ModelParams=WaveformPara(rfstream=rfstream, preonset=preonset, Sedthick=Sedthick, VpSed=VpSed, VpCrust=VpCrust, 
@@ -50,8 +56,11 @@ class TestWaveformFittingMultiproc(unittest.TestCase):
             gaussian=1.25
             savepath=save_tests()
             rayp=0.04
-            KMoho= np.linspace(1.65,1.95,121)
-            HMoho=np.linspace(20,60,201)
+            # KMoho= np.linspace(1.65,1.95,121)
+            # HMoho=np.linspace(20,60,201)
+            # Use smaller grid search discretization so the test runs faster
+            KMoho= np.linspace(1.65,1.95,10)
+            HMoho=np.linspace(25,50,10)
             nproc=25
             format='pdf'
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
