+++ docs/api.rst
@@ -1,0 +1,5 @@
+.. figure:: ./logo/rfsed_logo.png
+   :align: center
+
+.. automodule:: rfsed
+   :members:
