+++ README.md
@@ -27,8 +27,6 @@ Beside these receiver function methods, rfsed has the following features:
 Receiver functions streams in the rfsed are handled by the 'RFStream' class of the [rf](https://github.com/trichter/rf) open software, and it inherits a lot of useful methods from the Obspy class 'Stream'. It is supported via the obspyh5 package. For more information on class "RFStream", see documentation on [rf](https://rf.readthedocs.io/en/latest/). 
 The rf framework is included in the rfsed package for completness under The MIT License (MIT) Copyright (c) 2013-2019 Tom Eulenfeld. In the rfsed modules to extract earthquake data from local seismic records, read and write support for necessary metadata is provided for SAC, SeismicHanlder and HDF5 formats based on [ObsPy](https://github.com/obspy/obspy).
 
-The receiver function forward calculation in rfsed is done using SynR, which is modified after the [SynSeis module - seispy](https://github.com/xumi1993/seispy) project (under the GNU GENERAL PUBLIC LICENSE).
-
 
 
 
@@ -79,18 +77,20 @@ If you have any questions about the package, please use the [discussions feature
 All contributions are welcome ... e.g. report bugs, discuss or add new features.
 
 ## Citing rfsed
-If you found this package useful, please consider citing it.
-+ Akinremi S., van der Meijde, M., Thomas, C., Afonso, J. C., Ruigrok, E., & Fadel, I. (2024). Waveform fitting of receiver functions for enhanced retrieval of crustal structure in the presence of sediments. Journal of Geophysical Research: Solid Earth (under review)
+If you use `rfsed` in your work, please consider citing the following paper.
++ Akinremi S., van der Meijde, M., Thomas, C., Afonso, J. C., Ruigrok, E., & Fadel, I. (2024). Waveform fitting of receiver functions for enhanced retrieval of crustal structure in the presence of sediments. Journal of Geophysical Research: Solid Earth, 5(129). https://doi.org/10.1029/2023JB028393
 
 #### References
-+ Yeck, W. L., Sheehan, A. F., & Schulte-Pelkum, V. (2013). Sequential h-k stacking to obtain accurate crustal thicknesses beneath sedimentary basins. Bulletin of the Seismological Society of America, 103 , 2142-2150. doi: 10.1785/0120120290
-+ Yu, Y., Song, J., Liu, K. H., & Gao, S. S. (2015). Determining crustal structure beneath seismic stations overlying a low-velocity sedimentary layer using receiver functions. Journal of Geophysical Research: Solid Earth, 120 , 3208-3218. doi: 10.1002/2014JB011610
-+  Zhu, L., & Kanamori, H. (2000). Moho depth variation in southern California from teleseismic receiver functions. Journal of Geophysical Research: Solid Earth, 105, 2969-2980. doi: 10.1029/1999jb900322
++ Akinremi, S., van der Meijde, M., Thomas, C., Afonso J. C., Ruigrok E., Fadel, I. (2024). Waveform fitting of receiver functions for enhanced retrieval of crustal structure in the presence of sediments. Journal of Geophysical Research: Solid Earth, 5(129). https://doi.org/10.1029/2023JB028393
++ Tom Eulenfeld T., (2020). rf: Receiver function calculation in seismology. Journal of Open Source Software, 5(48), 1808, https://doi.org/10.21105/joss.01808
++ Yeck, W. L., Sheehan, A. F., & Schulte-Pelkum, V. (2013). Sequential h-k stacking to obtain accurate crustal thicknesses beneath sedimentary basins. Bulletin of the Seismological Society of America, 103 , 2142-2150. https://doi.org/10.1785/0120120290
[…]
