+++ docs/utils.rst
@@ -1,0 +1,8 @@
+.. figure:: ./logo/rfsed_logo.png
+   :align: center
+
+Module utils
+------------
+
+.. automodule:: telewavesim.utils
+   :members:
