+++ docs/elast.rst
@@ -1,0 +1,8 @@
+.. figure:: ./logo/rfsed_logo.png
+   :align: center
+
+Module elast
+------------
+
+.. automodule:: telewavesim.elast
+   :members:

[…]
