### 2024-05-25 0e63ae1
Add pytest to dependencies
