+++ setup.cfg
@@ -22,8 +22,9 @@ install_requires =
 	obspy>=1.0.3
 	rf>=1.0.3
 	matplotlib
-	numpy
+	numpy<=1.24.0
 	scipy
+	pytest>=7.4.4
 	colorama
 	OWSLib
 	plotly
@@ -32,7 +33,7 @@ install_requires =
 	pytz>=2020.1
 	importlib-metadata
 tests_require = 
-	pytest
+	pytest>=7.4.4
 	py
 zip_safe = False
 
