### 2024-05-25 05c7b06
Use smaller grid search discretization so the test runs faster
