+++ tests/test_WaveformFitting.py
@@ -21,8 +21,8 @@ class TestWaveformFitting(unittest.TestCase):
         # KMoho= np.linspace(1.65,1.95,121)
         # HMoho=np.linspace(20,60,201)
         # Use smaller grid search discretization so the test runs faster
-        KMoho= np.linspace(1.65,1.95,10)
-        HMoho=np.linspace(25,50,10)
+        KMoho= np.linspace(1.65,1.95,5)
+        HMoho=np.linspace(25,50,5)
         format='pdf'
         wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
         WaveformFitting(rfstream=rfstream, preonset=preonset, HSed=HSed, VpSed=VpSed, VpCrust=VpCrust, rayp=rayp, KMoho=KMoho, HMoho=HMoho, 
@@ -40,8 +40,8 @@ class TestWaveformFitting(unittest.TestCase):
             # KMoho= np.linspace(1.65,1.95,121)
             # HMoho=np.linspace(20,60,201)
             # Use smaller grid search discretization so the test runs faster
-            KMoho= np.linspace(1.65,1.95,10)
-            HMoho=np.linspace(25,50,10)
+            KMoho= np.linspace(1.65,1.95,5)
+            HMoho=np.linspace(25,50,5)
             wtCorr, wtRMSE, wtPG = [0.5, 0.3, 0.2] #wtCorr+wtRMSE+wtPG=1.0
             format='pdf'
             FittingResult=WaveformFitting(rfstream=rfstream, preonset=preonset, HSed=HSed, VpSed=VpSed, VpCrust=VpCrust, rayp=rayp, KMoho=KMoho, HMoho=HMoho, 
