### 2024-05-27 cf2d217
Change 'examples' directories locations
