+++ examples/notebooks/__init__.py
@@ -1,1 +1,0 @@
-# Examples of usage of rfsed
