### 2024-05-27 771cf82
Change 'examples' directories locations
