### 2024-05-27 3eace03
Update setup.cfg file
