+++ setup.cfg
@@ -21,10 +21,12 @@ python_requires = >=3.6, <3.11
 install_requires = 
 	obspy>=1.0.3
 	rf>=1.0.3
-	matplotlib
 	numpy<=1.24.0
-	scipy
+	obspyh5
+	matplotlib
 	pytest>=7.4.4
+	scikit-learn
+	seaborn
 	colorama
 	OWSLib
 	plotly
@@ -32,10 +34,16 @@ install_requires =
 	pyyaml
 	pytz>=2020.1
 	importlib-metadata
+	
+    
+	
 tests_require = 
 	pytest>=7.4.4
 	py
 zip_safe = False
+include_package_data = True
+
+
 
 [options.package_data]
 * = *.quakeml, *.dat, *.h5, *.jpg, *.pdf, *.xml, *.tar.gz, *.json, *.mseed, *.png
@@ -54,5 +62,3 @@ tests =
 
 [options.packages.find]
 where = src
-
-
