### 2024-05-27 e195c0e
implement 'data'and 'examples' directories change in util.py
