+++ src/rfsed/util.py
@@ -1,14 +1,14 @@
 from rf.rfstream import read_rf, rfstats
+from pkg_resources import resource_filename
 import numpy as np
 import os
-current_directory = os.path.dirname(os.path.abspath(__file__))
-main_dir = os.path.abspath(os.path.join(current_directory, '..', '..'))
+
 
 def rfMoho_example():
     """
     Return low-frequency receiver functions in a stream by read_rf().
     """
-    fname = main_dir + '/data/rfstreams_Moho/rfstreams.h5'
+    fname = resource_filename('rfsed', 'data/rfstreams_Moho/rfstreams.h5')
     stream = read_rf(fname)  
     return stream
 
@@ -16,7 +16,7 @@ def rfSed_example():
     """
     Return high-frequency receiver functions in a stream read_rf().
     """
-    fname = main_dir + '/data/rfstreams_Sed/rfstreams.h5'
+    fname = resource_filename('rfsed', 'data/rfstreams_Sed/rfstreams.h5')
     stream = read_rf(fname)  
     return stream
 
@@ -24,7 +24,7 @@ def raw_data_example():
     """
     Return directory to raw seismic record data.
     """
-    data_dir = main_dir + '/data/Raw_Data_DeepNL_Groningen/'
+    data_dir = resource_filename('rfsed', 'data/Raw_Data_DeepNL_Groningen/')
 
     return data_dir
 
@@ -32,14 +32,14 @@ def catalog_example():
     """
     Return directory to earthquake catalog.
     """
-    catalog_dir = main_dir + '/data/Earthquake_Catalog/eq_events.quakeml'
+    catalog_dir = resource_filename('rfsed', 'data/Earthquake_Catalog/eq_events.quakeml')
     return catalog_dir
 
 def read_raw_waveform_data():
     """
     Return streams of raw waeform data downloaded from IRIS using read_rf().
     """
-    fname = main_dir + '/examples/IRIS_Downloaded_Waveform/00_eq_data.h5'
+    fname = resource_filename('rfsed', 'examples/IRIS_Downloaded_Waveform/00_eq_data.h5')
     stream = read_rf(fname)  
     return stream
 
@@ -47,7 +47,7 @@ def save_Eq_data():
     """
     Return directory to save extracted earthquake data.
     """
-    savepath = main_dir +  '/examples/earthquake_data/'
+    savepath = resource_filename('rfsed', 'examples/earthquake_data/')
     if not os.path.exists(savepath):
         os.mkdir(savepath)
     return savepath
@@ -56,7 +56,7 @@ def save_IRIS_waveform():
     """
     Return directory to save downloaded earthquake waveform data.
     """
-    savepath = main_dir + '/examples/IRIS_Downloaded_Waveform/'
+    savepath = resource_filename('rfsed', 'examples/IRIS_Downloaded_Waveform/')
     if not os.path.exists(savepath):
         os.mkdir(savepath)
     return savepath
@@ -65,7 +65,7 @@ def save_calculated_RF():
     """
     Return directory to save calculated receiver function.
     """
-    savepath = main_dir + '/examples/Calculated_Receiver_Functions/'
+    savepath = resource_filename('rfsed', 'examples/Calculated_Receiver_Functions/')
     if not os.path.exists(savepath):
         os.mkdir(savepath)
     return savepath
@@ -74,7 +74,7 @@ def save_catalog():
     """
     Return directory to save extracted earthquake data.
     """
-    savepath = main_dir + '/examples/catalog/'
+    savepath = resource_filename('rfsed', 'examples/catalog/')
     if not os.path.exists(savepath):
         os.mkdir(savepath)
     return savepath
@@ -83,7 +83,7 @@ def save_plot():
     """
[…]
