### 2024-05-28 c81b6c7
Update the setup.cfg to include notebook and python files
