+++ setup.cfg
@@ -46,7 +46,7 @@ include_package_data = True
 
 
 [options.package_data]
-* = *.quakeml, *.dat, *.h5, *.jpg, *.pdf, *.xml, *.tar.gz, *.json, *.mseed, *.png
+* = *.quakeml, *.dat, *.h5, *.jpg, *.pdf, *.xml, *.tar.gz, *.json, *.mseed, *.png, *.ipynb, *.py
 
 
 [options.extras_require]
