### 2024-05-28 f59ab45
Updated util.py to include install_examples function
