### 2024-05-28 79fe718
Edited test_hkYu.py to exclude w1Sed, w2Sed, w3Sed parameters
