### 2024-05-28 791b68b
Extend the docstirng of the modules and follow the pep8 style guide
