+++ src/rfsed/SynthAnalysis.py
@@ -1,668 +1,0 @@
-"""
-.. module:: SynthAnalysis
-        :synopsis: Create synthetic reciever function data and analyse the data for hk, sequential hk, 
-        and resonance filtering
-.. moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
-"""
-
-import numpy as np
-from scipy.signal import detrend
-from scipy.fft import fft, ifft
-from scipy.signal import find_peaks
-import matplotlib.pyplot as plt
-from rfsed.synrf import synrf
-import matplotlib.gridspec as gridspec
-import matplotlib.transforms as mtransforms
-from matplotlib.legend_handler import HandlerLine2D as HL
-from scipy import signal
-from scipy.signal import argrelextrema, argrelmin, argrelmax
-import math
-import cmath
-#----------------------------------------------------------
-def getamp(rfdata, tarray, t):
-    """
-    Get the amplitude of the receiver function at a specific time
-
-    :param rfdata: receiver function data
-    :type rfdata: numpy array
-    :param tarray: time array
-    :type tarray: numpy array
-    :param t: time to get the amplitude
-    :type t: float
-    :return: amplitude of the receiver function at time t
-    """
-    amp = rfdata[(np.abs(tarray - t).argmin())]
-    return amp
-#----------------------------------------------------------
-def hkSynth(Synthrf, time, rayp=0.04, H=np.linspace(20,60,201), K=np.linspace(1.65,1.95,121), 
-            Vp=6.9, w1=0.6, w2=0.3, w3=0.1, layer = None):
-    """
-    Calculate the H-K stacking for synthetic receiver function after Zhu and Kanamori (2000)
-    Zhu, L., & Kanamori, H. (2000). Moho depth variation in southern california from
-    teleseismic receiver functions. Journal of Geophysical Research: Solid Earth, 105 , 
-    2969-2980. doi: 10.1029/1999jb900322
-    
-    :param Synthrf: synthetic receiver function
-    :type Synthrf: numpy array
-    :param time: time array
-    :type time: numpy array
-    :param rayp: ray parameter
-    :type rayp: float
-    :param H: depth array
-    :type H: numpy array
-    :param K: Vp/Vs array
-    :type K: numpy array
-    :param Vp: Vp value
-    :type Vp: float
-    :param w1: weight for Ps
-    :param w2: weight for PpPs
-    :param w3: weight for PsPs+PpSs
-    :type w1, w2, w3: float
-    :param layer: layer name 'Moho' or 'Sed'
-    :type layer: string
-    :return: Dictionary of H-K stacking result for synthetic receiver function
-    """
-    if Vp is None: Vp = 6.5
-    if w1 is None: w1 = 0.6
-    if w2 is None: w2 = 0.3
-    if w3 is None: w3 = 0.1   
-    if layer is None: layer = 'Moho'
-    #----------------------------------------------------------
-    rp=rayp
-    rfdata = Synthrf
-    t = time
-    stk = np.zeros((len(K)*len(H),3))
-    z = 0 
-    for i in range(len(K)):
-        Ktemp = K[i]
-        for j in range(len(H)):
-            Htemp = H[j]
-            s = 0.0
-            trdata = rfdata
-            #--------------------------------------
-            term1= ((Ktemp/Vp)**2 - rp**2)**0.5
[…]

[…]
