### 2024-05-28 fac0ece
Extend the docstirng of the modules and follow the pep8 style guide
