+++ src/rfsed/hkYu.py
@@ -1,0 +1,339 @@
+# Copyright (c) 2023, Stephen Akinremi
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+"""
+.. module:: hkYu
+        :synopsis: Modified H-K stacking method of Yu et al. (2015) for 
+        receiver functions filtered with the resonance filter
+        Yu, Y., Song, J., Liu, K. H., & Gao, S. S. (2015). Determining crustal 
+        structure beneath seismic stations overlying a low-velocity sedimentary 
+        layer using receiver functions. Journal of Geophysical Research: Solid 
+        Earth, 120 , 3208-3218. doi:10.1002/2014JB011610
+.. moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
+                  Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.gridspec as gridspec
+import matplotlib.transforms as mtransforms
+from matplotlib.legend_handler import HandlerLine2D as HL
+import seaborn
+seaborn.set_style("darkgrid", {"axes.facecolor": ".85"})
+
+#----------------------------------------------------------
+def getamp(rfdata, tarray, t):
+    """
+    Get the amplitude of the receiver function at a specific time
+
+    :param rfdata: receiver function data
+    :type rfdata: numpy array
+    :param tarray: time array
+    :type tarray: numpy array
+    :param t: time to get the amplitude
+    :type t: float
+    
+    Returns: 
+    Amplitude of the receiver function at time t
+    """
+    amp = rfdata[(np.abs(tarray - t).argmin())]
+    return amp
[…]
