+++ src/rfsed/ExtractEq.py
@@ -1,0 +1,143 @@
+# Copyright (c) 2023, Stephen Akinremi
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+"""
+.. module:: ExtractEq
+        :synopsis: Extract the earthquake waveform data  from the local data 
+                   files
+.. moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
+                  Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
+"""
+
+from obspy import read, read_inventory, Stream, read_events, UTCDateTime as UTC
+import numpy as np
+from obspy.core.event.catalog import read_events
+from glob import glob
+from obspy.geodetics.base import locations2degrees, gps2dist_azimuth
+from obspy.geodetics import kilometer2degrees
+from obspy.taup import TauPyModel
+#------------------------------------------------------------------------------
+def ExtractEq(datapath, filename, catalog, stalat, stalon, Request_window):
+        """
+        Extract the earthquake waveform data from the local data files
+        Writes the extracted data to a new file
+
+        :param datapath: Path to the local data files directory
+        :type datapath:str
+        :param filename: Path to the new file
+        :type filename:str
+        :param catalog: Path to the catalog file
+        :type catalog:str
+        :param stalat: Latitude of the station
+        :type stalat:float
+        :param stalon: Longitude of the station
+        :type stalon:float
+        :param Request_window: Time window relative to first P arrival 
+                               (in seconds)
+        :type Request_window:list[start, end]
+
+        Returns: 
+        Extracted earthquake data in a new file
+
+        Example
+        -------
+
+        Initialize the ExtractEq module:
+        >>> from rfsed.ExtractEq import ExtractEq
+        Define all the necessary parameters
[…]
