+++ docs/wiggles.rst
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo.png
-   :align: center
-
-Module wiggles
---------------
-
-.. automodule:: telewavesim.wiggle
-   :members:
