+++ docs/links.rst
@@ -1,17 +1,0 @@
-GitHub Repository
------------------
-
-* `rfsed <https://https://github.com/akinremisa/rfsed>`_
-
-
-References
-----------
-* Akinremi, S., van der Meijde, M., Thomas, C., Afonso J. C., Ruigrok E., Fadel, I. (2024). Waveform fitting of receiver functions for enhanced retrieval of crustal structure in the presence of sediments. Journal of Geophysical Research: Solid Earth, 5(129). https://doi.org/10.1029/2023JB028393
-
-* Tom Eulenfeld T., (2020). rf: Receiver function calculation in seismology. Journal of Open Source Software, 5(48), 1808, https://doi.org/10.21105/joss.01808
-  
-* Yeck, W. L., Sheehan, A. F., & Schulte-Pelkum, V. (2013). Sequential h-k stacking to obtain accurate crustal thicknesses beneath sedimentary basins. Bulletin of the Seismological Society of America, 103 , 2142-2150. https://doi.org/10.1785/0120120290
-
-* Yu, Y., Song, J., Liu, K. H., & Gao, S. S. (2015). Determining crustal structure beneath seismic stations overlying a low-velocity sedimentary layer using receiver functions. Journal of Geophysical Research: Solid Earth, 120 , 3208-3218. https://doi.org/10.1002/2014JB011610
-
-* Zhu, L., & Kanamori, H. (2000). Moho depth variation in southern California from teleseismic receiver functions. Journal of Geophysical Research: Solid Earth, 105, 2969-2980. https://doi.org/10.1029/1999jb900322
