+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. rfsed documentation master file, created by
+   sphinx-quickstart on Wed May 29 20:07:52 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to rfsed's documentation!
+=================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
