+++ docs/conf.py
@@ -1,28 +1,92 @@
+# # Configuration file for the Sphinx documentation builder.
+# #
+# # For the full list of built-in configuration values, see the documentation:
+# # https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# # -- Project information -----------------------------------------------------
+# # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+
+# project = 'rfsed'
+# copyright = '2024, Stephen Akinremi'
+# author = 'Stephen Akinremi'
+# release = '1.0.8'
+
+# # -- General configuration ---------------------------------------------------
+# # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
+
+# extensions = []
+
+# templates_path = ['_templates']
+# exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+
+
+
+# # -- Options for HTML output -------------------------------------------------
+# # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
+
+# html_theme = 'alabaster'
+# html_static_path = ['_static']
+
+
+
+
+
 # Configuration file for the Sphinx documentation builder.
 #
-# For the full list of built-in configuration values, see the documentation:
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 
-# -- Project information -----------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
+# -- Path setup --------------------------------------------------------------
 
-project = 'rfsed'
-copyright = '2024, Stephen Akinremi'
-author = 'Stephen Akinremi'
-release = '1.0.8'
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+#
+# import os
+# import sys
+# sys.path.insert(0, os.path.abspath('.'))
+
+
+# -- Project information -----------------------------------------------------
+
+project = 'telewavesim'
+copyright = '2019, Pascal Audet'
+author = 'Pascal Audet'
+
+# The full version, including alpha/beta/rc tags
+from telewavesim import __version__ as release
 
 # -- General configuration ---------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = []
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon']
 
-templates_path = ['_templates']
+autodoc_member_order = 'bysource'
+
+html_logo = '../telewavesim/examples/picture/logo_tws.png'
+
+# Add any paths that contain templates here, relative to this directory.
+#templates_path = ['_templates']
+
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
-
+# The name of the Pygments (syntax highlighting) style to use.
+pygments_style = 'sphinx'
 
 # -- Options for HTML output -------------------------------------------------
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'alabaster'
-html_static_path = ['_static']
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+#
+html_theme = 'sphinx_rtd_theme'
+
[…]
