+++ docs/.travis.yml
@@ -1,7 +1,7 @@
 deploy:
   provider: pages
   skip_cleanup: true
-  github_token: $GITHUB_TOKEN  # Set in the settings page of your repository, as a secure variable
+  github_token: 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  # Set in the settings page of your repository, as a secure variable
   keep_history: true
   on:
     branch: main
