+++ docs/.travis.yml
@@ -4,4 +4,4 @@ deploy:
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
   on:
-    branch: main
+    branch: gh-pages
