+++ docs/requirements.txt
@@ -1,5 +1,0 @@
-# requirements to build the docs
-alabaster
-decorator
-obspy
-sphinx
