+++ .travis.yml
@@ -2,9 +2,10 @@ language: minimal
 
 
 deploy:
-  provider: pages
+  provider: pages:git
   skip_cleanup: true
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
+  edge: true
   on:
-    branch: gh-pages
+    branch: main
