+++ .travis.yml
@@ -7,5 +7,6 @@ deploy:
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
   edge: true
+  local_dir: docs/_build
   on:
     branch: main
