+++ .travis.yml
@@ -6,7 +6,6 @@ deploy:
   skip_cleanup: true
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
-  edge: true
-  local_dir: docs/_build
+  local_dir: build
   on:
     branch: main
