+++ .travis.yml
@@ -1,11 +1,9 @@
-language: minimal
-
 
 deploy:
   provider: pages:git
   skip_cleanup: true
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
-  local_dir: build
+
   on:
     branch: main
