+++ .travis.yml
@@ -1,9 +1,7 @@
-
 deploy:
-  provider: pages:git
+  provider: pages
   skip_cleanup: true
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
-
   on:
     branch: main
