+++ .travis.yml
@@ -1,3 +1,6 @@
+language: minimal
+
+
 deploy:
   provider: pages
   skip_cleanup: true
