+++ .travis.yml
@@ -6,5 +6,6 @@ deploy:
   skip_cleanup: true
   github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
   keep_history: true
+  local_dir: build
   on:
     branch: gh-pages
