+++ .travis.yml
@@ -8,4 +8,4 @@ deploy:
   keep_history: true
   local_dir: build
   on:
-    branch: gh-pages
+    branch: main
