+++ docs/conf.py
@@ -14,15 +14,20 @@ release = '1.0.8'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = []
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon']
 
-templates_path = ['_templates']
+autodoc_member_order = 'bysource'
+
+html_logo = './logo/rfsed_logo.png'
+
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
-
+# The name of the Pygments (syntax highlighting) style to use.
+pygments_style = 'sphinx'
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'alabaster'
-html_static_path = ['_static']
+html_theme = 'sphinx_rtd_theme'
+# html_theme = 'alabaster'
+# html_static_path = ['_static']
