+++ docs/_sources/api.rst.txt
@@ -1,0 +1,5 @@
+.. figure:: ./logo/rfsed_logo_horizontal.png
+   :align: center
+
+.. automodule:: rfsed
+   :members:
