+++ docs/api.rst
@@ -1,5 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-.. automodule:: rfsed
-   :members:
