### 2024-05-30 07a16ff
delete docs/
