+++ docs/apiuyuiyui.rst
@@ -1,0 +1,5 @@
+.. figure:: ./logo/rfsed_logo_name.png
+   :align: center
+
+.. automodule:: rfsed
+   :members:
