+++ docs/api.rst
@@ -1,0 +1,2 @@
+.. figure:: ./logo/rfsed_logo_name.png
+   :align: center
