### 2024-05-31 0004dfe
delete docs/
