+++ docs/ExtractEqMultiproc.rst
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module ExtractEqMultiproc
-------------
-
-.. automodule:: rfsed.ExtractEqMultiproc
-   :members:
