+++ docs/_sources/gettingstarted.rst.txt
@@ -1,0 +1,6 @@
+Documentation
+================
+
+rfsed is developed specifically to implement different techniques of **analysing receiver functions
+from stations overlying sedimentary layers.** The package is **adaptable, efficient, and easy-to-use** for different analysis receiver functions
+obtained from stations overlying sedimentary layers.
