### 2024-05-31 cd1e7bb
delete docs/
