+++ README.md
@@ -6,25 +6,34 @@
 [![python version](https://img.shields.io/pypi/pyversions/rf.svg)](https://python.org)
 
 
-rfsed is developed specifically to implement different techniques of **analysing receiver functions
-from stations overlying sedimentary layers.** The software is **adaptable, efficient, and easy-to-use** for different analysis receiver functions
-obtained from stations overlying sedimentary layers.
+**rfsed** is developed specifically to implement different techniques of 
+**analysing receiver functions from stations overlying sedimentary layers.** 
+The software is **adaptable, efficient, and easy-to-use** for different 
+analysis of receiver functions obtained from stations overlying sedimentary layers.
 
-Receiver functions techniques implemented in rfsed are:
-+ H-k stacking of Zhu and Kanamori (2000)
-+ Sequential H-k stacking of Yeck et al., (2013)
+Receiver functions techniques implemented in **rfsed** are:
++ H-k stacking (one layer) of Zhu and Kanamori (2000)
++ Sequential H-k stacking (two layers) of Yeck et al., (2013)
 + Resonance filtering and modified H-k stacking of Yu et al., (2015)
-+ H-k stacking and Waveform Fitting two-step method of Akinremi et al., (2024)
++ H-k stacking and Waveform Fitting (two-step method) of Akinremi et al., (2024)
 + Analysis of the synthetic reciever functions with the above-mentioned methods.
 
 
-Beside these receiver function methods, rfsed has the following features:
-+ Creating publication quality figures for the results of the analysis.
+Beside these receiver function methods, **rfsed** has the following features:
 + Extracting earthquake data from local seismic record files.
-+ Multiprocessing options for waveform fitting and extracting earthquake data from local seismic record files.
++ Multiprocessing options for waveform fitting and extracting earthquake 
+data from local seismic record files.
++ Creating publication quality figures for the results of the analysis.
 
 
-Receiver functions streams in the rfsed are handled by the 'RFStream' class of the [rf](https://github.com/trichter/rf) open software, and it inherits a lot of useful methods from the Obspy class 'Stream'. It is supported via the obspyh5 package. For more information on class "RFStream", see documentation on [rf](https://rf.readthedocs.io/en/latest/). In the rfsed modules to extract earthquake data from local seismic records, read and write support for necessary metadata is provided for SAC, SeismicHanlder and HDF5 formats based on [ObsPy](https://github.com/obspy/obspy).
+Receiver functions streams in the **rfsed** are handled by the 'RFStream' class of the 
+[rf](https://github.com/trichter/rf) open software, and it inherits a lot of useful 
+methods from the Obspy class 'Stream'. It is supported via the obspyh5 package. 
+For more information on class "RFStream", see documentation on 
+[rf](https://rf.readthedocs.io/en/latest/). In the **rfsed** modules to extract earthquake 
+data from local seismic records, read and write support for necessary metadata is 
+provided for SAC, SeismicHanlder and HDF5 formats based on 
+[ObsPy](https://github.com/obspy/obspy).
 
 
 
@@ -32,16 +41,17 @@ Receiver functions streams in the rfsed are handled by the 'RFStream' class of t
 ## Installation and testing of this software
 
 ### Installation from PyPi
-The easiest way to install rfsed is via `pip`::
+The easiest way to install **rfsed** is via `pip`::
 
 ```bash
 pip install rfsed
 ```
 
 ### Installation development version from source code
[…]
