+++ README.md
@@ -26,9 +26,9 @@ data from local seismic record files.
 + Creating publication quality figures for the results of the analysis.
 
 
-Receiver functions streams in the **rfsed** are handled by the 'RFStream' class of the 
+Receiver functions streams in the **rfsed** are handled by the `RFStream` class of the 
 [rf](https://github.com/trichter/rf) open software, and it inherits a lot of useful 
-methods from the Obspy class 'Stream'. It is supported via the obspyh5 package. 
+methods from the Obspy class `Stream`. It is supported via the obspyh5 package. 
 For more information on class "RFStream", see documentation on 
 [rf](https://rf.readthedocs.io/en/latest/). In the **rfsed** modules to extract earthquake 
 data from local seismic records, read and write support for necessary metadata is 
@@ -110,7 +110,7 @@ Research: Solid Earth, 120 , 3208-3218. https://doi.org/10.1002/2014JB011610
 + Zhu, L., & Kanamori, H. (2000). Moho depth variation in southern California from teleseismic 
 receiver functions. Journal of Geophysical Research: Solid Earth, 105, 2969-2980. https://doi.org/10.1029/1999jb900322
 
-##### Related receiver function projects
+#### Related receiver function projects
 + [rf](https://github.com/trichter/rf) including calculation of receiver functions
 + [seispy](https://github.com/xumi1993/seispy) including hk-stacking
 + [RFPy](https://github.com/paudetseis/RfPy) including hk-stacking, harmonic decomposition
