+++ docs/_sources/WaveformFitting.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module WaveformFitting
-------------
-
-.. automodule:: rfsed.WaveformFitting
-   :members:
