+++ docs/ReverbFilter.rst
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module ReverbFilter
-------------
-
-.. automodule:: rfsed.ReverbFilter
-   :members:
