+++ docs/_sources/hkZhu.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module hkZhu
-------------
-
-.. automodule:: rfsed.hkZhu
-   :members:
