+++ docs/hkSeqYeck.rst
@@ -1,0 +1,8 @@
+.. figure:: ./logo/rfsed_logo_horizontal.png
+   :align: center
+
+Module hkSeqYeck
+-----------------
+
+.. automodule:: rfsed.hkSeqYeck
+   :members:
