### 2024-05-31 993d6f4
delete docs/
