### 2024-05-31 55f1945
Modified module docstring
