+++ src/rfsed/WaveformFittingMultiproc.py
@@ -25,13 +25,12 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-"""
-..module:: WaveformFittingMultiproc
-        :synopsis: Calculate the Waveform Fitting for each Moho depth 
-                   in parallel
-.. moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
-                  Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
-"""
+#
+# module:: WaveformFittingMultiproc
+#      :synopsis: Calculate the Waveform Fitting for each Moho depth in parallel
+# moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
+#                Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
+# """
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -262,6 +261,7 @@ def run_waveformfitting(nproc, HMoho, ModelParams):
     pool.join()
     return WaveformFitResults
 #----------------------------------------------------------
+
 def plotbestmodel(WaveformResults, ModelParams, wtCorr, wtRMSE, wtPG, 
                   savepath, format):
     """
