+++ src/rfsed/SynthAnalysis.py
@@ -25,13 +25,13 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-"""
-.. module:: SynthAnalysis
-        :synopsis: Create synthetic reciever function data and analyse the 
-                data for hk, sequential hk and resonance filtering
-.. moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
-                  Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
-"""
+# 
+# module:: SynthAnalysis
+#      :synopsis: Create synthetic reciever function data and analyse the 
+#                 data for hk, sequential hk and resonance filtering
+# moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
+#                Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
+
 
 import numpy as np
 from scipy.signal import detrend
@@ -62,9 +62,7 @@ def getamp(rfdata, tarray, t):
     return amp
 #----------------------------------------------------------
 
-def hkSynth(Synthrf, time, rayp=0.04, H=np.linspace(20,60,201), 
-            K=np.linspace(1.65,1.95,121), Vp=6.9, w1=0.6, w2=0.3, w3=0.1,
-            layer = None):
+def hkSynth(Synthrf, time, rayp, H, K, Vp, w1, w2, w3, layer = None):
     """
     Calculate the H-K stacking for synthetic receiver function after 
     Zhu and Kanamori (2000)
@@ -189,11 +187,8 @@ def hkSynth(Synthrf, time, rayp=0.04, H=np.linspace(20,60,201),
     return HKResultSynth
 #----------------------------------------------------------
 
-def hkSeqSynth(rfSynthSed, rfSynthMoho, time, rayp=0.04, 
-               HSed=np.linspace(0,10,201), KSed=np.linspace(1.65,2.25,201),
-               VpSed=3.0, w1Sed=0.6, w2Sed=0.3, w3Sed=0.1, 
-               HMoho=np.linspace(20,60,201), KMoho=np.linspace(1.65,1.95,121),
-               VpMoho=6.9, w1Moho=0.6, w2Moho=0.3, w3Moho=0.1, 
+def hkSeqSynth(rfSynthSed, rfSynthMoho, time, rayp, HSed, KSed, VpSed, w1Sed, 
+               w2Sed, w3Sed, HMoho, KMoho, VpMoho, w1Moho, w2Moho, w3Moho, 
                g = [75.,10., 15., 2.5], rmneg = None):
     """
     Calculate the Sequential H-K stacking for synthetic receiver 
