### 2024-05-31 da9faff
Modified module docstring
