### 2024-05-31 5c9e47f
Modified module docstring
