### 2024-06-01 4499d9d
Modified module docstring
