+++ src/rfsed/ExtractEqMultiproc.py
@@ -57,26 +57,27 @@ def Add_sta_time_2cat(catalogfile, stalat, stalon, Request_window):
                            (time before P arrival, time after P arrival)
     :type Request_window: list [start, end]
 
+
     Returns:
-    Catalog of events with station latitude, longitude and request time 
-    window added to each event
+        Catalog of events with station latitude, longitude and request time 
+        window added to each event
 
     Example
     --------
 
-    Initialize the Add_sta_time_2cat function:
+    >>> # Initialize the Add_sta_time_2cat function:
     >>> from rfsed.ExtractEqMultiproc import Add_sta_time_2cat
-    Define the input parameters:
+    >>> # Define the input parameters:
     >>> catalogfile = 'path/to/catalogfile'
     >>> stalat = 52.0
     >>> stalon = 4.0
     >>> Request_window = [-50, 150]
-    Add the station latitude, longitude and request time window to the 
-    catalog using the Add_sta_time_2cat function:
+    >>> # Add the station latitude, longitude and request time window to the 
+    >>> # catalog using the Add_sta_time_2cat function:
     >>> catalog = Add_sta_time_2cat(catalogfile, stalat, stalon,
                                     Request_window)
-
     """
+
     catalog = read_events(catalogfile)
     for i in catalog:
         i.stalat = stalat
@@ -97,17 +98,18 @@ def Eq_times(catalog):
     :type catalog: obspy.core.event.catalog.Catalog
 
     Returns: 
-    Dictionary of Earthquake start and end time for the requested time window 
-    relative to the first P arrival
+        Dictionary of Earthquake start and end time for the requested time window 
+        relative to the first P arrival
 
     Example
     --------
-    Initialize the Eq_times function:
+
+    >>> # Initialize the Eq_times function:
     >>> from rfsed.ExtractEqMultiproc import Eq_times
-    Define the input parameters (output from the Add_cat_sta_time function):
+    >>> # Define the input parameters (output from the Add_cat_sta_time function):
     >>> catalog = Add_cat_sta_time(catalogfile, stalat, stalon, Request_window)
-    Get the earthquake start and end times for the requested time window 
-    relative to the first P arrival using the Eq_times function:
+    >>> # Get the earthquake start and end times for the requested time window 
+    >>> # relative to the first P arrival using the Eq_times function:
     >>> timewindow = Eq_times(catalog)
 
     """
@@ -146,22 +148,25 @@ def Get_Eqtimes_multiproc(catalog, nproc):
     :param nproc: Number of processors to use
     :type nproc: int
 
-    Returns: List of dictionaries of Earthquake start and end times 
-    (in obspy UTCDateTime) for the requested time window relative to the first P arrival
+    
+    Returns: 
+        List of dictionaries of Earthquake start and end times 
+        (in obspy UTCDateTime) for the requested time window relative to the first P arrival
 
     Example
     --------
 
-    Initialize the Get_Eqtimes_multiproc function:
+    >>> # Initialize the Get_Eqtimes_multiproc function:
     >>> from rfsed.ExtractEqMultiproc import Get_Eqtimes_multiproc
-    Define the input parameters catalog and nproc (number of processors to use)
-    The catalog is an output from the Add_cat_sta_time function
+    >>> # Define the input parameters catalog and nproc (number of processors to use)
+    >>> # he catalog is an output from the Add_cat_sta_time function
     >>> catalog = Add_cat_sta_time(catalogfile, stalat, stalon, Request_window)
     >>> nproc = 4
-    Get the earthquake start and end times for the requested time window 
[…]
