### 2024-06-01 08cb5e1
Modified module docstring
