### 2024-06-01 edea0bd
delete docs/
