+++ docs/ExtractEq.rst
@@ -1,0 +1,8 @@
+.. figure:: ./logo/rfsed_logo_horizontal.png
+   :align: center
+
+Module ExtractEq
+------------------
+
+.. automodule:: rfsed.ExtractEq
+   :members:
