### 2024-06-01 d34720e
modified waveform fitting discretization
