### 2024-06-01 ed30aa8
delete docs/
