### 2024-06-01 1f93ff3
delete docs/
