+++ docs/_sources/WaveformFittingMultiproc.rst.txt
@@ -1,0 +1,8 @@
+.. figure:: ./logo/rfsed_logo_horizontal.png
+   :align: center
+
+Module WaveformFittingMultiproc
+----------------------------------
+
+.. automodule:: rfsed.WaveformFittingMultiproc
+   :members:
