+++ src/rfsed/__init__.py
@@ -1,104 +1,33 @@
-# Copyright (c) 2023 Stephen Akinremi MIT license
+# Copyright (c) 2023 Stephen Akinremi BSD 3-Clause License
 """
 rfsed Documentation
 ================
+**rfsed** is developed specifically to implement different techniques of 
+**analysing receiver functions from stations overlying sedimentary layers.** 
+The software is **adaptable, efficient, and easy-to-use** for different 
+analysis of receiver functions obtained from stations overlying sedimentary layers.
 
-rfsed is a Python framework for analysizing receiver functions.
-It is developed specifically to implement different techniques of analysing receiver functions
-from stations overlying sedimentary layers. Receiver functions techniques implemented in rfsed are:
-1. H-k stacking of Zhu and Kanamori (2000)
-2. Sequential H-k stacking of Yeck et al., (2013)
-3. Resonance filtering and modified H-k stacking of Yu et al., (2015)
-4. H-k stacking and Waveform Fitting two-step method of Akinremi et al., (2024)
-6. Analysis of the synthetic reciever functions with the above-mentioned methods.
-
-Beside these receiver function methods, rfsed has the following features:
-1. Creating publication quality figures for the results of the analysis.
-2. Extracting earthquake data from local seismic record files.
-3. Multiprocessing options for waveform fitting and extracting earthquake data from local seismic record files.
+Receiver functions techniques implemented in **rfsed** are:
++ H-k stacking (one layer) of Zhu and Kanamori (2000)
++ Sequential H-k stacking (two layers) of Yeck et al., (2013)
++ Resonance filtering and modified H-k stacking of Yu et al., (2015)
++ H-k stacking and Waveform Fitting (two-step method) of Akinremi et al., (2024)
++ Analysis of the synthetic reciever functions with the above-mentioned methods.
 
 
-Receiver functions streams are handled by the 'RFStream' class of the rf open software, and it inherits a lot of useful methods from the Obspy class 'Stream'.
-It is supported via the obspyh5 package. For more information on class "RFStream", see documentation on rf (https://rf.readthedocs.io/en/latest/). 
-The rf framework is included in the rfsed package for completness under The MIT License (MIT) Copyright (c) 2013-2019 Tom Eulenfeld.
-In the module to extract earthquake data from local seismic records, read and write support for necessary metadata is provided 
-for SAC, SeismicHanlder and HDF5 formats based on ObsPy.
-
-The forward receiver function calculation is done using SynR, which is modified after the SynSeis module of Mijian Xu's seispy project (under the GNU GENERAL PUBLIC LICENSE).
-
-Method
-------
-The receiver function technique is a well-established technique to image velocity contrasts in the subsurface by using isolated P–to-S wave 
-conversions (P-receiver functions) and the reverberations generated at such discontinuities. The velocity contrast at the crust-mantle 
-boundary (Moho) represents a major discontinuity that receiver functions are highly sensitive to, making them useful to constrain average 
-crustal thickness and Vp/Vs. Several methods such as H-k stacking, has been developed to investigate the crustal thickness and Vp/Vs ratio 
-using receiver functions. However, the presence of near-surface low-velocity sedimentary layer can obscure Moho signals due to the additional 
-P-to-S wave conversions and  associated reverberated phases created at the sediment-basement discontinuity. These additional intra-crustal phases 
-can have large amplitudes and similar arrival times to Moho signals, which makes it difficult to retrieve Moho information using standard RF techniques.
-Several other methods have been developed to deal with the complex RF in sedimentary basins and recover crustal strucures (crustal thickness and Vp/Vs).
[…]
