+++ src/rfsed/examples/test_output/__init__.py
@@ -1,1 +1,1 @@
-# Examples of usage of rfsed
+# Test outputs
