+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = rfsed
-version = 0.1.8
+version = 0.1.9
 author = Stephen Akinremi
 author_email = s.akinremi@utwente.nl
 description = Package for receiver functions analysis and dealing with sediment effects in receiver functions
