### 2024-06-01 akinremisa/rfsed#2 @akinremisa
Hi @paudetseis

Thank you for your review of our repository. Your comments are very helpful in improving the accessibility of the software.

1.	A local installation of the notebook examples is added after installation of **rfsed** via **PyPI**. The notebooks are now more detailed step-by-step of how to use the modules.
2.	The environment.yml has been excluded from the repository. And a guide on how to create an environment and install the packages is described in the documentation.
3.	A proper and extended documentation is now done for **rfsed**. Although the html files are still in the repository
4.	The docstring for the modules are now extended to include what the function does, the parameters, how to initialize/import the function, and a simple demonstration of how to use it. This is also included in the documentation.
5.	The readability of the code has been improved using **pep8**
