### 2024-06-01 akinremisa/rfsed#3 @akinremisa
Hi @alistairboyce11

Thank you for your review of our software.
1.	We have now reduced the discretization for the grid search waveform fitting process in the test to quicken it. This now runs faster.
2.	The environment.yml is now excluded from the repository and other smooth installation guides provided.
3.	The notebooks are now more described extensively step-by-step to help users find the code more accessible. The discretization for the grid search waveform fitting process in the examples is reduced to quicken it
