+++ .travis.yml
@@ -1,10 +1,0 @@
-language: minimal
-
-
-deploy:
-  provider: pages
-  skip_cleanup: true
-  github_token: $GITHUB_TOKEN # Set in the settings page of your repository, as a secure variable###
-  keep_history: true
-  on:
-    branch: gh-pages
