### 2024-06-03 beccb1c
delete docs/
