+++ docs/_sources/hkYu.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module hkYu
-------------
-
-.. automodule:: rfsed.hkYu
-   :members:
