+++ docs/_sources/hkSeqYeck.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module hkSeqYeck
------------------
-
-.. automodule:: rfsed.hkSeqYeck
-   :members:
