### 2024-06-03 8eebad3
delete docs/
