+++ docs/_sources/ExtractEq.rst.txt
@@ -1,8 +1,0 @@
-.. figure:: ./logo/rfsed_logo_horizontal.png
-   :align: center
-
-Module ExtractEq
-------------------
-
-.. automodule:: rfsed.ExtractEq
-   :members:
