+++ joss/paper.md
@@ -27,28 +27,28 @@ bibliography: paper.bib
 
 # Introduction
 
-The receiver function technique is a well-established technique to image velocity contrasts in the subsurface (such as the crust-mantle boundary - Moho) by using isolated P–to-S wave conversions (P-receiver functions) and the reverberations generated at such discontinuities. Several methods such as H-k stacking [@Zhu2000], have been developed to investigate the average crustal thickness and Vp/Vs ratio using receiver functions. However, the presence of a near-surface low-velocity sedimentary layer can obscure Moho phases due to the additional P-to-S wave conversions and associated reverberated phases created at the sediment-basement discontinuity. These additional intra-crustal phases can have large amplitudes and similar arrival times as the Moho phases, which makes it difficult to retrieve Moho information using standard RF techniques. 
+The receiver function technique is a well-established technique to image velocity contrasts in the subsurface (such as the crust-mantle boundary - Moho) using isolated P–to-S wave conversions (P-receiver functions) and the reverberations generated at such discontinuities. Several methods such as H-k stacking [@Zhu2000], have been developed to investigate the average crustal thickness and Vp/Vs ratio using receiver functions. However, the presence of a near-surface low-velocity sedimentary layer can obscure Moho phases due to the additional P-to-S wave conversions and associated reverberated phases created at the sediment-basement discontinuity. These additional intra-crustal phases can have large amplitudes and similar arrival times as the Moho phases, which makes it difficult to retrieve Moho information using standard receiver function techniques. 
 
 # Statement of need
-``rfsed`` is a Python package for receiver function analysis that includes methods for dealing with sediment effects. ``rfsed`` presents a new approach to retrieving the sediment thickness properties using H-K stacking of the high-frequency receiver function, followed by a waveform fitting approach to retrieve the average crustal thickness and Vp/Vs [@Akinremi2024]. Moreover, ``rfsed`` contains implementations of the most common receiver function approaches for dealing with the sediment effect with possible synthetic testing capabilities. ``rfsed`` has already been used in [@Akinremi2024] for multimethod imaging of the crustal structure using receiver functions obtained from stations overlying seimentary layer.
+``rfsed`` is a Python software for receiver function analysis that includes methods for dealing with sediment effects. ``rfsed`` presents a new approach for retrieving reliable crustal thickness and Vp/Vs from stations overlying sedimentary layer. The techniques derives sediment thickness and Vp/Vs using H-K stacking of the high-frequency receiver function, followed by a waveform fitting approach to retrieve the average crustal thickness and Vp/Vs [@Akinremi2024]. Moreover, ``rfsed`` contains implementations of the most common receiver function approaches for dealing with the sediment effect with possible synthetic testing capabilities. ``rfsed`` has already been used in [@Akinremi2024] for multimethods imaging of the crustal structure using receiver functions obtained from stations overlying sedimentary layer.
 
 # Key Functionality
[…]
