### 2024-06-17 akinremisa/rfsed#3 @alistairboyce11
Hi @akinremisa,

Thank you for making the above changes. They should help ensure the rfsed package is well used in the community.

Best wishes,

Alistair
