### 2024-06-24 akinremisa/rfsed#2 @paudetseis
Thank you, @akinremisa, for making those revisions to the software. I can now install `rfsed` without a glitch using both `pip` and cloning. The documentation is also much improved, and the html doc is highly appreciated. I still have a few suggestions for improving the user's experience:

- [ ] The automated tests are appreciated (although they produce weird results, probably because they are meant to quickly test the code), but the corresponding documentation is still a bit misleading. Simply typing `pytest` after installation will produce very different results whether the software is installed via `pip` or through cloning, because the source code is not in the current directory when installed via `pip`. I suggest you either tell the user where to go to run the tests depending on the install method, or you specify that tests are only available for those who clone the code.
- [ ] The installation of the notebooks works as advertised in the documentation. However, when I tried running Example1 in Jupyter, I got a message saying that `tqdm` is unavailable (not installed). This should be fixed in the installation (setup.cfg).
- [ ] Some notebooks (e.g., Example6) have the output of the last jupyter run saved and displayed. I suggest saving "clean" versions for all notebooks (the successfully executed versions are online anyways).
- [ ] I am a bit worried about the quality of the notebook results. For example, Examples 6 to 8 show a poor estimation of Vp/Vs and Depth, where the best-fit solutions are at the edges of the model. None of them appear to be very satisfactory. Is this on purpose? Maybe to show that the waveform fitting is superior? I am surprised to see that most methods fail to estimate model parameters for this station. It is not a great selling point of the software if the examples don't give any decent results. Can you either change the example data or be more explicit in the notebook about these data and what you observe and interpret from the results? Describe the experiment in more detail and the motivation for choosing poor examples? Or, I'm guessing, the example data are the same as the ones used in testing, and are much more limited than a typical dataset?
- [ ] When I started paying attention to the actual waveform data used in the examples, I noticed a mix-up in stations between Example 1 (station HGN) and Examples 3-5 (station NE301) and then Examples 6-8,11-12 (station OPLO). This is quite confusing to the user, and I recommend cleaning this up to offer a complete analysis for a single station.
