### 2024-07-02 #6612 @elbeejay
@akinremisa at this time I believe we are waiting on you to make some revisions or respond to the concerns in https://github.com/akinremisa/rfsed/issues/2, @paudetseis has indicated where he believes the revisions fall a bit short of what he was looking to see.
