### 2024-07-03 #6612 @akinremisa
@paudetseis Thank you for your new comments and review. I am now working on fixing those. @elbeejay I am working on the revisions from [Comment1](https://github.com/akinremisa/rfsed/issues/2) and will revert back
