### 2024-07-08 04aed72
Added tqdm as dependency in setup.cfg file
