+++ setup.cfg
@@ -24,6 +24,7 @@ install_requires =
 	numpy<=1.24.0
 	obspyh5
 	matplotlib
+	tqdm
 	pytest>=7.4.4
 	scikit-learn
 	seaborn
