### 2024-07-15 3afda2c
Modified appending onset attribute based on number of traces
