+++ src/rfsed/data/rfstreams_Moho/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2023 Stephen Akinremi MIT license
-"""
-Data examples for testing.
-
-"""
