+++ src/rfsed/ExtractEqMultiproc.py
@@ -1,0 +1,269 @@
+# Copyright (c) 2023, Stephen Akinremi
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+# 
+# module:: ExtractEqMultiproc
+#      :synopsis: Extract the data for each earthquake from the local 
+#                    data files in parallel
+# moduleauthor:: Stephen Akinremi <s.akinremi@utwente.nl> 
+#                   Islam Fadel <i.e.a.m.fadel@utwente.nl> (October 2023)
+#
+
+from obspy import read, read_events, Stream
+import numpy as np
+from obspy.core.event.catalog import read_events
+from obspy.geodetics.base import locations2degrees, gps2dist_azimuth
+from obspy.geodetics import kilometer2degrees
+from obspy.taup import TauPyModel
+import multiprocessing
+import itertools
+
+def Add_sta_time_2cat(catalogfile, stalat, stalon, Request_window):
+    """
+    Add station latitude, longitude and request time window to each 
+    event in the catalog
+
+    :param catalogfile: Path to the catalog file
+    :type catalogfile: str
+    :param stalat: Station latitude
+    :type stalat: float
+    :param stalon: Station longitude
+    :type stalon: float
+    :param Request_window: Request time window relative to first P arrival 
+                           (time before P arrival, time after P arrival)
+    :type Request_window: list [start, end]
+
+
+    Returns:
+        Catalog of events with station latitude, longitude and request time 
+        window added to each event
+
+    Example
+    --------
+
+    >>> # Initialize the Add_sta_time_2cat function:
+    >>> from rfsed.ExtractEqMultiproc import Add_sta_time_2cat
+    >>> # Define the input parameters:
+    >>> catalogfile = 'path/to/catalogfile'
+    >>> stalat = 52.0
[…]
