+++ src/rfsed/util.py
@@ -1,0 +1,261 @@
+# Copyright (c) 2023, Stephen Akinremi
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+from rf.rfstream import read_rf
+from pkg_resources import resource_filename
+import os
+import pkg_resources as _pkg_resources
+from distutils import dir_util as _dir_util
+
+def rfMoho_example():
+    """
+    Return low-frequency receiver functions in a stream by read_rf().
+
+    Example
+    -------
+
+    >>> # Initialize the rfMoho_example module:
+    >>> from rfsed.util import rfMoho_example
+    >>> rfMoho=rfMoho_example()
+    """
+
+    fname = resource_filename('rfsed', 'data/rfstreams_Moho/rfstreams.h5')
+    stream = read_rf(fname)  
+    return stream
+
+def rfSed_example():
+    """
+    Return high-frequency receiver functions in a stream read_rf().
+    
+    Example
+    -------
+
+    >>> # Initialize the rfSed_example module:
+    >>> from rfsed.util import rfSed_example
+    >>> rfSed=rfSed_example()
+    """
+
+    fname = resource_filename('rfsed', 'data/rfstreams_Sed/rfstreams.h5')
+    stream = read_rf(fname)  
+    return stream
+
+def raw_data_example():
+    """
+    Return directory to raw seismic record data.
+
+    Example
+    -------
+
+    >>> # Initialize the raw_data_example module:
+    >>> from rfsed.util import raw_data_example
+    >>> data_dir=raw_data_example()
+    """
+
+    data_dir = resource_filename('rfsed', 'data/Raw_Data_DeepNL_Groningen/')
+    # data_dir = resource_filename('rfsed', 'data/Local_Raw_Data/')
+
+    return data_dir
+
+def catalog_example():
+    """
[…]
