+++ src/rfsed/data/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2023 Stephen Akinremi MIT license
+"""
+Data examples for testing.
+
+"""
