+++ src/rfsed/examples/test_output/__init__.py
@@ -1,0 +1,1 @@
+# Test outputs
