+++ src/rfsed/examples/plots/__init__.py
@@ -1,0 +1,1 @@
+# Examples of usage of rfsed
