+++ src/rfsed/examples/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) 2023 Stephen Akinremi MIT license
+"""
+Examples
+
+"""
