### 2024-07-22 661f4c7
Add option to install test files locally
