### 2024-07-22 7294236
add rfsed_tests to examples
