+++ src/rfsed/examples/rfsed_tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Test rfsed 
+"""

[…]
