+++ docs/_sources/ReverbFilter.rst.txt
@@ -1,0 +1,8 @@
+.. figure:: ./logo/rfsed_logo_horizontal.png
+   :align: center
+
+Module ReverbFilter
+--------------------
+
+.. automodule:: rfsed.ReverbFilter
+   :members:
