### 2024-07-22 akinremisa/rfsed#2 @akinremisa
Hi @paudetseis

Thank you for the new comments. I now addressed them in the repository and also provide the details of the change made below.

1. The results of the automated tests (waveform fitting) look as such because they use smaller discretization to make them run faster. On running the tests, this is now more clearly described in the documentation. The pytest can be used in the package directory when rfsed is installed by cloning the repository. There is an option to install the automated files locally and then run the tests. This is described in the documentation.

2. 'tqdm' is a dependency of rf and should be installed automatically along side other dependencies. However, it is also now included in the setup.cfg file.

3. Because I use Jupyter notebook viewer that is directly linked to the example files in the repository, the same version (executed) is kept for both the online viewing and the repository itself.

4. I now use a single station data for all the examples (Station OPLO). The station is overlying sedimentary rock and the estimation of the Moho depth and Vp/Vs is poorly constrained when using one-layer H-k stacking, Sequential H-k stacking and the resonance filtering methods. This effect of sediment on the estimation in the examples is not clearly described in the notebook. Also, the waveform fitting approach estimate a more reliable Moho depth and Vp/Vs estimates.

5. Now, same data (single station - OPLO) is used for all the examples and tests.
