### 2024-08-03 #6612 @elbeejay
Checking in here, @paudetseis it looks like @akinremisa has made revisions and replied to your comments in https://github.com/akinremisa/rfsed/issues/2. Do you feel your comments have been sufficiently addressed?

If so, if you could check off the last few items in your reviewer checklist above that'd be great.
