### 2024-08-06 akinremisa/rfsed#2 @paudetseis
Tahnk you @akinremisa
I appreciate these changes and can now recommend publication at JOSS. I will now close this comment and mark it as complete. Good work!
