### 2024-08-09 #6612 @elbeejay
@paudetseis - do you mind checking off the last few boxes in your reviewer checklist above? I understand that your comments were resolved in https://github.com/akinremisa/rfsed/issues/2, but we like to ask our reviewers to formally complete their checklists for posterity in the event we ever had to go back through JOSS reviews.
