### 2024-08-14 #6612 @elbeejay
@akinremisa please let me know what the latest version number is, share the DOI of the archive once it is released on Zenodo, figshare, or an equivalent, and just check the metadata as indicated above. In the meantime I'll take care of those editor tasks and will let you know if I have any last comments or requests.
