### 2024-08-18 de11c47
Edittd manuscript
