### 2024-08-20 4e86d50
Added jpg Figures in the Manuscript
