### 2024-08-20 #6612 @akinremisa
> @akinremisa I'd ask that you consider:
>
> * increasing the font size on figure labels; the figures render correctly, but the labels are a bit tough to read without zooming in quite a bit on the PDF
> * Line 20: replace "derives" with the singular "derive"

@elbeejay Thank you for your suggested edits. I have implemented these suggestions in the manuscript.
