### 2024-08-20 #6612 @akinremisa
> @akinremisa please let me know what the latest version number is, share the DOI of the archive once it is released on Zenodo, figshare, or an equivalent, and just check the metadata as indicated above. In the meantime I'll take care of those editor tasks and will let you know if I have any last comments or requests.

@elbeejay

The latest version of the software is **1.0.0**

I have also uploaded this latest version to Zenodo.

Here is the Zenodo DOI: https://doi.org/10.5281/zenodo.13348322
