### 2024-08-21 #6612 @elbeejay
@akinremisa can you please revise the title of the zenodo archive to match the paper exactly? Right now it is "rfsed: Receiver functions analysis and dealing with sediment effects in receiver functions." while the paper is titled: "rfsed: Receiver function analysis and dealing with sediment effects"

Once that is done I can recommend that we accept the paper and publish it in JOSS.
